% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_estimators_Pauly.R
\name{B0}
\alias{B0}
\title{trace estimator B_0}
\usage{
B0(X)
}
\arguments{
\item{X}{data matrix with subjects in cols and factor levels in rows multiplied with the hypothesis matrix}
}
\value{
returns the trace estimator for T*V
}
\description{
trace estimator B_0
}
\keyword{internal}
