% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_estimators_Pauly.R
\name{B2}
\alias{B2}
\title{trace estimator B_2}
\usage{
B2(X)
}
\arguments{
\item{X}{data matrix with subjects in cols and factor levels in rows multiplied with the hypothesis matrix}
}
\value{
returns the trace estimator for (T*V)^2
}
\description{
trace estimator B_2
}
\keyword{internal}
