% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_estimators_Pauly.R
\name{B3}
\alias{B3}
\title{trace estimator B_3}
\usage{
B3(X)
}
\arguments{
\item{X}{data matrix with subjects in colss and factor levels in rows multiplied with the hypothesis matrix T}
}
\value{
returns the trace estimator for (T*V)^3
}
\description{
trace estimator B_3
}
\keyword{internal}
