% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-glmnet.R
\name{boot.glmnet}
\alias{boot.glmnet}
\title{Calculate confidence intervals for lasso using bootstrap}
\usage{
boot.glmnet(X, y, B = 500, lambda, seed, alpha = 0.05, ...)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response vector}

\item{B}{Number of bootstrap resamples (default 500)}

\item{lambda}{Regularization parameter at which solutions are to be bootstrapped (by default, uses cross-valiation to find lambda)}

\item{seed}{Seed (optional)}

\item{alpha}{Error rate; 0.05 (default) corresponds to 95% confidence interval}
}
\description{
Calculate confidence intervals for lasso using bootstrap
}
