% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.r
\name{cache}
\alias{cache}
\title{Evaluate expression and cache, or load from cache}
\usage{
cache(expr, path)
}
\arguments{
\item{expr}{An R expression that returns a \code{data.frame}-compatible object;
see \verb{[arrow::write_parquet()]}}

\item{path}{A file path (where to save/load the object)}
}
\description{
If the file exists, the expression is not evaluated -- the cached result is
returned. If the file path does not exist, the expression is evaluated and
stored as a parquet file at the location specified by the path. To trigger a
re-run, delete the file.
}
\examples{
tmp <- tempfile()
cache(Ex1.1(N=10), tmp)  # saves to cache (slow)
cache(Ex1.1(N=10), tmp)  # loads from cache (fast)
}
