% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_utils.R
\name{check_criteria_grouped}
\alias{check_criteria_grouped}
\title{Checks whether everything is as expected for the grouped test}
\usage{
check_criteria_grouped(X, group, hypothesis, reps, subsampling)
}
\arguments{
\item{X}{data matrix which is investigated}

\item{group}{vector containing the allocation to the
subjects of the measurement}

\item{hypothesis}{a hypothesis given through a character to specify a predefined
standard hypothesis or by a list}

\item{reps}{a \code{string} determines the number of subsamples used by the
subsampling trace estimators}

\item{subsampling}{a logical number specifying whether subsampling versions of
estimators should be used if existing}
}
\value{
an error message if any condition is violated
}
\description{
This function ensures that all criteria are checked in one place,
}
\keyword{internal}
