% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_utils.R
\name{check_criteria_single}
\alias{check_criteria_single}
\title{Checks whether everything is as expected for the single test}
\usage{
check_criteria_single(X, hypothesis)
}
\arguments{
\item{X}{data matrix which is investigated}

\item{hypothesis}{a hypothesis given through a character to specify a predefined
standard hypothesis or by a list}
}
\value{
an error message if any condition is violated
}
\description{
This function ensures that all criteria are checked in one place,
}
\keyword{internal}
