% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIplot.R
\name{CIplot}
\alias{CIplot}
\alias{CIplot.matrix}
\title{Plot confidence intervals (NEEDS DOCUMENTATION)}
\usage{
CIplot(obj, ...)

\method{CIplot}{matrix}(
  obj,
  labels = rownames(B),
  sort = TRUE,
  pxlim,
  xlim,
  ylim,
  sub,
  diff = (ncol(B) == 4),
  null = 0,
  n.ticks = 6,
  mar,
  axis = !add,
  trans,
  p.label = FALSE,
  xlab = "",
  ylab = "",
  add = FALSE,
  setupOnly = FALSE,
  lwd = 2,
  replaceUnderscore = TRUE,
  ...
)
}
\arguments{
\item{obj}{The object to be plotted; can be a matrix of raw values or a model object}
}
\description{
"Forest plot"-style plotting of confidence intervals from a regression model.  Basic input is a matrix with columns of estimate/lower/upper, along with an optional 4th column for the p-value.  Also works with a variety of models (lm/glm/coxph/etc.)
}
\examples{
# Supplying a matrix
B <- cbind(1:3, 0:2, 2:4)
rownames(B) <- LETTERS[1:3]
CIplot(B)

# Supplying a fitted model object
fit <- lm(Ozone ~ Solar.R + Wind + Temp, airquality)
CIplot(fit)

# Options

}
