% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{download_data}
\alias{download_data}
\alias{read_data}
\alias{attach_data}
\title{Download data and read into R}
\usage{
download_data(name)

read_data(name)

attach_data(name, envir = parent.frame())
}
\arguments{
\item{name}{Name of data set to read/attach}

\item{envir}{If attaching, to which environment?}
}
\description{
These functions are downloading, reading, and attaching data sets described in the hdrm book.
To keep the size of the \code{hdrm} package manageable, these data sets are not included with the package and must be downloaded separately using \code{downloadData()}; this only needs to be done once.
}
\details{
Data sets in hdrm are lists consisting of \code{X}, \code{y}, and possibly other elements.
\code{read_data} returns the list as an object that the user can name and use how they want.
\code{attach_data} assigns the elements of the list into the user's environment.
\code{attach_data} is often more convenient, but offers less control and potentially introduces side effects, such as overwriting an existing object \code{X} if one already exists.  See the examples for more details.
}
\examples{
\dontrun{
download_data()         # Download all data sets
download_data(brca1)    # Download a specific data set}
Data <- read_data(brca1)
dim(Data$X)
head(Data$y)
attach_data(brca1)
dim(X)
head(y)
}
