% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{downloadData}
\alias{downloadData}
\alias{readData}
\alias{attachData}
\title{Download data and read into R}
\usage{
downloadData(name)

readData(name)

attachData(name, envir = parent.frame())
}
\arguments{
\item{name}{Name of data set to read/attach}

\item{envir}{If attaching, to which environment?}
}
\description{
These functions are downloading, reading, and attaching data sets described in the hdrm book.
To keep the size of the `hdrm` package manageable, these data sets are not included with the package and must be downloaded separately using `downloadData()`; this only needs to be done once.
}
\details{
Data sets in hdrm are lists consisting of `X`, `y`, and possibly other elements.
`readData` returns the list as an object that the user can name and use how they want.
`attachData` assigns the elements of the list into the user's environment.
`attachData` is often more convenient, but offers less control and potentially introduces side effects, such as overwriting an existing object `X` if one already exists.  See the examples for more details.
}
\examples{
\dontrun{
downloadData()         # Download all data sets
downloadData(bcTCGA)   # Download a specific data set}


Data <- readData(bcTCGA)
dim(Data$X)
head(Data$y)


attachData(bcTCGA)
dim(X)
head(y)

}
