\name{Ex1.1}
\alias{Ex1.1}
\title{Reproduce Example 1.1}
\usage{
Ex1.1(N = 100, n = 25, p = 100)
}
\description{
\code{Ex1.1()} will recreate Example 1.1 from the book.  If you specify any
options, your results may look different.  See the code for details.
}
\arguments{
  \item{N}{Number of simulated realizations}

  \item{n}{Sample size}

  \item{p}{Number of features}

}
\examples{
Ex1.1()
}
