% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam-snr.r
\name{Ex3.1}
\alias{Ex3.1}
\title{Reproduce Example 3.1}
\usage{
Ex3.1(N = 500, n = 50, p = 100, p1 = 6, SNR = c(1, 2, 4), seed = 2)
}
\arguments{
\item{N}{Number of simulated realizations}

\item{n}{Sample size}

\item{p}{Number of features}

\item{p1}{Number of non-null features}

\item{SNR}{Vector of SNR values}

\item{seed}{Seed for reproducibility}
}
\value{
A list of two arrays, \code{MCP}, containing the MSE for MCP, and \code{SCAD}, containing the MSE for SCAD.
}
\description{
Reproduces Example 3.1 from the book.  If you specify any options, your results may look different.
}
\examples{
Ex3.1(N=2)

}
