% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enet-sim.R
\name{Ex4.2}
\alias{Ex4.2}
\title{Reproduce Example 4.2}
\usage{
Ex4.2(
  N = 1000,
  rho = seq(0, 0.9, 0.1),
  n = 50,
  p = 100,
  p1 = 5,
  b = 0.5,
  alpha = 0.5,
  corr = c("cs", "bd"),
  seed = 1
)
}
\arguments{
\item{N}{Number of simulated realizations}

\item{rho}{Vector of correlations}

\item{n}{Sample size}

\item{p}{Number of features}

\item{p1}{Number of non-null features}

\item{b}{Coefficient associated with the non-null features}

\item{alpha}{Elastic net tuning parameter}

\item{corr}{Correlation structure, either \code{cs} (compound symmetric) or \code{bd} (block diagonal)}

\item{seed}{Seed for reproducibility}
}
\description{
Reproduces Example 4.2 from the book.  If you specify any options, your results may look different.
}
\examples{
Ex4.2(N=10)

}
