% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet-sim-ind.R
\name{Ex4.3}
\alias{Ex4.3}
\title{Reproduce Example 4.3}
\usage{
Ex4.3(N = 100, s = seq(0.1, 1.1, 0.2), n = 100, p = 500, p1 = 12, seed = 1)
}
\arguments{
\item{N}{Number of simulated realizations}

\item{s}{Signal strength (coefficient value for the non-null features)}

\item{n}{Sample size}

\item{p}{Number of features}

\item{p1}{Number of non-null features}

\item{seed}{Seed for reproducibility}
}
\description{
Reproduces Example 4.3 from the book.  If you specify any options, your results may look different.
}
\examples{
Ex4.3(N=10)
}
