% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdr-ind.r
\name{Ex6.2}
\alias{Ex6.2}
\alias{Fig6.3}
\title{Reproduce Example 6.2 and Figure 6.3}
\usage{
Ex6.2()

Fig6.3(out)
}
\arguments{
\item{out}{Output of \code{Ex6.2()}}
}
\description{
Reproduces Example 6.2 and Figure 6.3 from the book.  If you specify any options, your results may look different.
}
\examples{
\dontrun{
out <- Ex6.2()
Fig6.3(out)
}
}
