% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdr-corr.R
\name{Ex6.3}
\alias{Ex6.3}
\alias{Fig6.5}
\title{Reproduce Example 6.3 and Figure 6.5}
\usage{
Ex6.3()

Fig6.5(out)
}
\arguments{
\item{out}{Output of \code{Ex6.3()}}
}
\description{
Reproduces Example 6.3 and Figure 6.5 from the book.  If you specify any options, your results may look different.
}
\examples{
\dontrun{
out <- Ex6.3()
Fig6.5(out)
}
}
