% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling-simdata.R
\name{Ex9.1}
\alias{Ex9.1}
\title{Reproduce Example 9.1}
\usage{
Ex9.1(n = 100, p = 60, a = 6, b = 2, rho = 0.5, beta = c(1, -1, 0.5,
  -0.5, 0.5, -0.5), seed = 1, ...)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of features}

\item{a}{Number of causal ('A') variables}

\item{b}{Number of correlated ('B') variables}

\item{rho}{Correlation between 'A' and 'B' variables}

\item{beta}{Coefficients for causal variables}

\item{seed}{Random number seed for reproducibility}

\item{...}{Further arguments to \code{\link{genDataABN}}}
}
\description{
Reproduces Example 9.1 from the book.  If you specify any options, your results may look different.
}
\examples{
out <- Ex9.1()
}
