% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-tcga.R
\name{Ex9.2}
\alias{Ex9.2}
\title{Reproduce Example 9.2}
\usage{
Ex9.2(N = 100)
}
\arguments{
\item{N}{Number of random splits}
}
\description{
Reproduces Example 9.2 from the book.  If you specify any options, your results may be different.
}
\examples{
p <- Ex9.2(N=10)
head(sort(p))
}
