\name{Fig1.1}
\alias{Fig1.1}
\title{Reproduce Figure 1.1}
\usage{
Fig1.1(n = 20, p = n - 1, seed = 1)
}
\description{
\code{Fig1.1()} will recreate Figure 1.1 from the book.  If you specify any
options, your results may look different.  See the code for details.
}
\arguments{
  \item{n}{Sample size}

  \item{p}{Number of features}

  \item{seed}{Random seed}

}
\examples{
Fig1.1()
}
