\name{Fig1.2}
\alias{Fig1.2}
\title{Reproduce Figure 1.2}
\usage{
Fig1.2(out)
}
\description{
\code{Fig1.2()} will recreate Figure 1.2 from the book.  If you specify any
options, your results may look different.  See the code for details.
}
\arguments{
  \item{out}{Output of Example}

}
\examples{
Fig1.2()
}
