% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse-ridge.r
\name{Fig1.3}
\alias{Fig1.3}
\title{Reproduce Figure 1.3}
\usage{
Fig1.3(rho = 0.5, n = 20)
}
\arguments{
\item{rho}{Correlation}

\item{n}{a}
}
\description{
Reproduces Figure 1.3 from the book.  If you specify any options, your results may look different.
}
\examples{
Fig1.3()

}
