% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-trace.r
\name{Fig1.4}
\alias{Fig1.4}
\title{Reproduce Figure 1.4}
\usage{
Fig1.4(loocv = TRUE, mar = c(5, 5, 5, 7))
}
\arguments{
\item{loocv}{Include a vertical line at the lambda value that minimizes LOOCV? (Default: true)}

\item{mar}{Margins (passed to par)}
}
\description{
Reproduces Figure 1.4 from the book.  If you specify any options, your results may look different.
}
\examples{
Fig1.4()
}
