% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{Fig2.1}
\alias{Fig2.1}
\title{Reproduce Figure 2.1}
\usage{
Fig2.1(
  range = c(-5, 5),
  col = c("#FF4E37FF", "#008DFFFF"),
  parlist = list(mfrow = c(1, 2), mar = c(4, 4, 0.5, 0.5), mgp = c(2, 1, 0), oma = c(0,
    0, 2, 0))
)
}
\arguments{
\item{range}{Range for beta coefficient (vector of length 2)}

\item{col}{Lasso/ridge color (vector of length 2)}

\item{parlist}{List of arguments to pass to `par()`}
}
\description{
Reproduces Figure 2.1 from the book; if you specify any options, your results may look different.
}
\examples{
Fig2.1()
}
