% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-tcga.r
\name{Fig2.10}
\alias{Fig2.10}
\title{Reproduce Figure 2.10}
\usage{
Fig2.10(cvfit, which = c("both", "left", "right"), parlist)
}
\arguments{
\item{cvfit}{\code{glmnet()} fit to the TCGA data; see examples}

\item{which}{\code{left}, \code{right}, or \code{both}}

\item{parlist}{List of arguments to pass to \code{par()}}
}
\description{
Reproduces Figure 2.10 from the book.  If you specify any options, your results may look different.
}
\examples{
attach_data(brca1)
cvfit <- cv.glmnet(X, y)
Fig2.10(cvfit)
Fig2.10(cvfit, 'left')
Fig2.10(cvfit, 'right')
}
