% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-tcga-ic.r
\name{Fig2.11}
\alias{Fig2.11}
\title{Reproduce Figure 2.11}
\usage{
Fig2.11(fit)
}
\arguments{
\item{fit}{\code{ncvreg()} fit to the TCGA data; see examples}
}
\description{
Reproduces Figure 2.11 from the book.  If you specify any options, your results may look different.
}
\examples{
attachData(bcTCGA)
fit <- ncvreg(X, y, penalty='lasso', lambda.min=0.005)
Fig2.11(fit)

}
