% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbotax.r
\name{Fig2.12}
\alias{Fig2.12}
\title{Reproduce Figure 2.12}
\usage{
Fig2.12(cvfit, parlist = list(mfrow = c(1, 2), mar = c(5, 5, 5, 0.5)))
}
\arguments{
\item{cvfit}{\code{cv.ncvreg()} fit to the TCGA data; see examples}

\item{parlist}{List of arguments to pass to \code{par()}}
}
\description{
Reproduces Figure 2.12 from the book.  If you specify any options, your results may look different.
}
\examples{
# Set up data
attach_data(Koussounadis2014)
sDay <- splines::ns(sData$Day, df=2)
X0 <- model.matrix(~ Treatment*sDay, sData)[,-1]
multiplier <- rep(0:1, c(ncol(X0), ncol(X)))
XX <- cbind(X0, X)

# Fit
fold <- c(4, 3, 10, 7, 3, 2, 5, 9, 6, 9, 7, 4, 4, 2, 1, 7, 1, 4, 6, 9, 5, 9, 9,
          6, 4, 8, 2, 2, 5, 5, 1, 1, 5, 7, 9, 9, 8, 7, 4, 5,  4, 2, 8, 9, 3, 1,
          8, 8, 10, 10, 3, 8, 3, 3, 9, 5, 1, 10, 10, 2, 7, 2, 4, 4, 8, 7, 6, 6,
          7, 10, 3, 6, 10, 2, 7, 5, 2, 6, 6, 1, 1, 4, 6, 10, 7, 8, 10, 8, 1, 5,
          5, 10, 6, 10, 3, 1, 3, 8, 2, 9, 3)
cvfit <- cv.ncvreg(XX, y, penalty.factor=multiplier, penalty='lasso', fold=fold)

# Plot
Fig2.12(cvfit)

}
