% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresh.R
\name{Fig2.2}
\alias{Fig2.2}
\title{Reproduce Figure 2.2}
\usage{
Fig2.2(range = c(-2.5, 2.5), col = c("#FF4E37FF", "#008DFFFF"))
}
\arguments{
\item{range}{Range for beta coefficient (vector of length 2)}

\item{col}{Lasso/ridge color (vector of length 2)}
}
\description{
Reproduces Figure 2.2 from the book; if you specify any options, your results may look different.
}
\examples{
Fig2.2()

}
