% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ-samp-dist.R
\name{Fig2.3}
\alias{Fig2.3}
\title{Reproduce Figure 2.3}
\usage{
Fig2.3(lambda = 0.5, n = 10, sigma = 1, ymax = 1.5)
}
\arguments{
\item{lambda}{Regularization parameter}

\item{n}{Sample size}

\item{sigma}{Standard deviation of noise}

\item{ymax}{What density should Pr(1) correspond to?}
}
\description{
Reproduces Figure 2.3 from the book; if you specify any options, your results may look different.
}
\examples{
Fig2.3()

}
