% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic-lasso.r
\name{Fig2.6}
\alias{Fig2.6}
\title{Reproduce Figure 2.6}
\usage{
Fig2.6(EBIC = FALSE)
}
\arguments{
\item{EBIC}{Include the extended BIC (EBIC, Chen & Chen 2008) in the figure?}
}
\description{
Reproduces Figure 2.6 from the book; if you specify any options, your results may look different.
}
\examples{
Fig2.6()
Fig2.6(EBIC=TRUE)

}
