% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvlasso.r
\name{Fig2.7}
\alias{Fig2.7}
\title{Reproduce Figure 2.7}
\usage{
Fig2.7()
}
\description{
Reproduces Figure 2.7 from the book; if you specify any options, your results may look different.
}
\examples{
Fig2.7()

# Default plot
attach_data(pollution)
cvfit <- cv.glmnet(std(X), y)   # Standardize design matrix
plot(cvfit)

# Range of lambda values within 1 SE of minimum
lmin <- which(cvfit$lambda==cvfit$lambda.min)
range(cvfit$lambda[which(cvfit$cvm < cvfit$cvup[lmin])])

}
