% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.r
\name{Fig2.8}
\alias{Fig2.8}
\title{Reproduce Figure 2.8}
\usage{
Fig2.8(
  n = 100,
  p = 1000,
  N = 25,
  parlist = list(mfrow = c(1, 2), mar = c(5, 5, 0.5, 0.5), oma = c(0, 0, 2, 0)),
  seed = 1
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of features}

\item{N}{Number of splits for the refitted CV estimator}

\item{parlist}{List of arguments to pass to \code{par()}}

\item{seed}{Random number seed for reproducibility}
}
\description{
Reproduces Figure 2.8 from the book; if you specify any options, your results may look different.
}
\examples{
Fig2.8(N=5)  # Increase N for a more reliable estimate
}
