% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsq.r
\name{Fig2.9}
\alias{Fig2.9}
\title{Reproduce Figure 2.9}
\usage{
Fig2.9()
}
\description{
Reproduces Figure 2.9 from the book; if you specify any options, your results may look different.
}
\examples{
Fig2.9()

# By hand
attachData(pollution)
cvfit <- cv.glmnet(std(X), y)   # Standardize design matrix
rsq <- 1-cvfit$cvm/var(y)

}
