% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp-scad-al.r
\name{Fig3.2}
\alias{Fig3.2}
\title{Reproduce Figure 3.2}
\usage{
Fig3.2(
  n = 200,
  p = 1000,
  seed = 105,
  ylim = c(-4.1, 4.1),
  parlist = list(mfrow = c(2, 2), mar = c(4.5, 4.5, 3, 0.5))
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of features}

\item{seed}{Random number seed for reproducibility}

\item{ylim}{Vertical limits, passed to plot()}

\item{parlist}{List of arguments to pass to \code{par()}}
}
\description{
Reproduces Figure 3.2 from the book; if you specify any options, your results may look different.
}
\examples{
Fig3.2()
}
