% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias-variance-gamma.R
\name{Fig3.3}
\alias{Fig3.3}
\title{Reproduce Figure 3.3}
\usage{
Fig3.3(N = 1e+05, parlist = list(mfrow = c(1, 2), mar = c(5, 5, 1.5,
  0.5), oma = c(0, 0, 2, 0)))
}
\arguments{
\item{N}{Number of simulations}

\item{parlist}{List of arguments to pass to `par()`}
}
\description{
Reproduces Figure 3.3 from the book; if you specify any options, your results may look different.
}
\examples{
Fig3.3()
}
