% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam-effect.r
\name{Fig3.4}
\alias{Fig3.4}
\title{Reproduce Figure 3.4}
\usage{
Fig3.4(
  n = 200,
  p = 1000,
  seed = 105,
  ylim = c(-4, 4),
  parlist = list(mfrow = c(1, 3), mar = c(4, 4, 2, 0), cex = 1)
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of features}

\item{seed}{Random number seed for reproducibility}

\item{ylim}{Vertical limits, passed to plot()}

\item{parlist}{List of arguments to pass to \code{par()}}
}
\description{
Reproduces Figure 3.4 from the book; if you specify any options, your results may look different.
}
\examples{
Fig3.4()
}
