% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-tcga-al.r
\name{Fig3.7}
\alias{Fig3.7}
\title{Reproduce Figure 3.7}
\usage{
Fig3.7(seed = 1, nfolds = 10, ...)
}
\arguments{
\item{seed}{For reproducibility}

\item{nfolds}{As in \code{cv.ncvreg()} (default: 10)}

\item{...}{Additional arguments to \code{plot.cv.ncvreg()}}
}
\description{
Reproduces Figure 3.7 from the book.  If you specify any options, your results may look different.
}
\examples{
Fig3.7(nfolds=3)
}
