% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-tcga-scad.R
\name{Fig3.9}
\alias{Fig3.9}
\title{Reproduce Figure 3.9}
\usage{
Fig3.9(gam = 8, seed = 1, parlist = list(mfrow = c(1, 2)))
}
\arguments{
\item{gam}{Gamma parameter (default: 8)}

\item{seed}{For reproducibility}

\item{parlist}{List of arguments to pass to `par()`}
}
\description{
Reproduces Figure 3.9 from the book.  If you specify any options, your results may look different.
}
\examples{
Fig3.9()
}
