% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enet-sim.R
\name{Fig4.1}
\alias{Fig4.1}
\title{Reproduce Figure 4.1}
\usage{
Fig4.1(
  cs,
  bd,
  parlist = list(mfrow = c(1, 2), mar = c(5, 5, 0.5, 0.5), oma = c(0, 0, 2, 0))
)
}
\arguments{
\item{cs}{Output from Ex4.2 with compound symmetric structure}

\item{bd}{Output from Ex4.2 with block diagonal structure}

\item{parlist}{List of arguments to pass to `par()`}
}
\description{
Reproduces Figure 4.1 from the book.  If you specify any options, your results may look different.
}
\examples{
cs <- Ex4.2(N=5, corr='cs')
bd <- Ex4.2(N=5, corr='bd')
Fig4.1(cs, bd)
}
