% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet-sim-ind.R
\name{Fig4.2}
\alias{Fig4.2}
\title{Reproduce Figure 4.2}
\usage{
Fig4.2(res, parlist = list(mfrow = c(1, 2), mar = c(4, 4, 2, 0.5)))
}
\arguments{
\item{parlist}{List of arguments to pass to `par()`}

\item{cs}{Output from Ex4.3 with compound symmetric structure}

\item{bd}{Output from Ex4.3 with block diagonal structure}
}
\description{
Reproduces Figure 4.2 from the book.  If you specify any options, your results may look different.
}
\examples{
cs <- Ex4.3(N=5, corr='cs')
bd <- Ex4.3(N=5, corr='bd')
Fig4.2(cs, bd)
}
