% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet-sim-cor.R
\name{Fig4.3}
\alias{Fig4.3}
\title{Reproduce Figure 4.3}
\usage{
Fig4.3(res, parlist = list(mfrow = c(1, 2), mar = c(4, 4, 2, 0.5)))
}
\arguments{
\item{res}{Output from Ex4.4}

\item{parlist}{List of arguments to pass to `par()`}
}
\description{
Reproduces Figure 4.3 from the book.  If you specify any options, your results may look different.
}
\examples{
res <- Ex4.4(N=3)
Fig4.3(res)

}
