% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-tcga-mnet.R
\name{Fig4.4}
\alias{Fig4.4}
\title{Reproduce Figure 4.4}
\usage{
Fig4.4(
  alpha = c(0.75, 0.5, 0.25),
  ylim = c(-0.25, 0.5),
  seed = 1,
  nfolds = 10,
  parlist = list(mfrow = c(1, 3), mar = c(4, 4, 2, 0), cex = 1)
)
}
\arguments{
\item{alpha}{Vector of alpha values for MNet penalty; there will be one plot per alpha value}

\item{ylim}{Vertical limits of plots}

\item{seed}{For reproducibility}

\item{nfolds}{As in \code{cv.ncvreg()}}

\item{parlist}{List of arguments to pass to \code{par()}}
}
\description{
Reproduces Figure 4.4 from the book.  If you specify any options, your results may look different.
}
\examples{
Fig4.4(nfolds=3)
}
