\name{Fig5.3}
\alias{Fig5.3}
\title{Reproduce Figure 5.3}
\usage{
Fig5.3(out)
}
\description{
\code{Fig5.3()} will recreate Figure 5.3 from the book.  If you specify any
options, your results may look different.  See the code for details.
}
\arguments{
  \item{out}{Output of Example}

}
\examples{
Fig5.3()
}
