\name{Fig5.4}
\alias{Fig5.4}
\title{Reproduce Figure 5.4}
\usage{
Fig5.4(out)
}
\description{
\code{Fig5.4()} will recreate Figure 5.4 from the book.  If you specify any
options, your results may look different.  See the code for details.
}
\arguments{
  \item{out}{Output of Example}

}
\examples{
Fig5.4()
}
