\name{Fig5.5}
\alias{Fig5.5}
\title{Reproduce Figure 5.5}
\usage{
Fig5.5(out)
}
\description{
\code{Fig5.5()} will recreate Figure 5.5 from the book.  If you specify any
options, your results may look different.  See the code for details.
}
\arguments{
  \item{out}{Output of Example}

}
\examples{
Fig5.5()
}
