% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-values.R
\name{Fig6.1}
\alias{Fig6.1}
\title{Reproduce Figure 6.1}
\usage{
Fig6.1(out)
}
\arguments{
\item{out}{Output of Ex6.1()}
}
\description{
Reproduces Figure 6.1 from the book; if you specify any options, your results may look different.
}
\examples{
out <- Ex6.1()
Fig6.1(out)

}
