% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdr-power.r
\name{Fig6.4}
\alias{Fig6.4}
\title{Reproduce Figure 6.4}
\usage{
Fig6.4(out)
}
\arguments{
\item{out}{Output of \code{Ex6.2()}}
}
\description{
Reproduces Example Figure 6.4 from the book.  If you specify any options, your results may look different.
}
\examples{
\dontrun{
out <- Ex6.2()
Fig6.4(out)
}
}
