% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdr-tcga.r
\name{Fig6.6}
\alias{Fig6.6}
\title{Reproduce Figure 6.6}
\usage{
Fig6.6(fit)
}
\arguments{
\item{fit}{\code{ncvreg} fit to the TCGA data; see examples}
}
\description{
Reproduces Figure 6.6 from the book.  If you specify any options, your results may look different.
}
\examples{
attachData(brca1)

fit <- ncvreg(X, y, penalty="lasso")
obj <- mfdr(fit)
tail(subset(obj, mFDR < .1))
plot(obj)

# Lasso figure
Fig6.6(fit)

# An MCP figure
fit <- ncvreg(X, y)
Fig6.6(fit)

}
