% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-split.R
\name{Fig9.1}
\alias{Fig9.1}
\title{Reproduce Figure 6.6}
\usage{
Fig9.1(res, N = 100, seed = 1)
}
\arguments{
\item{N}{Number of random splits}

\item{seed}{Random number seed for reproducibility}

\item{out}{Output of Ex9.1()}
}
\description{
Reproduces Figure 6.6 from the book.  If you specify any options, your results may look different.
}
\examples{

Data <- Ex9.1()
res <- Fig9.1(Data)
}
