% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stab-toy.R
\name{Fig9.2}
\alias{Fig9.2}
\title{Reproduce Figure 9.2}
\usage{
Fig9.2(out, N = 100, seed = 1)
}
\arguments{
\item{out}{Output of Ex9.1}

\item{N}{Number of simulated realizations}

\item{seed}{Random number seed for reproducibility}
}
\description{
Reproduces Figure 9.2 from the book.  If you specify any options, your results may look different.
}
\examples{
out <- Ex9.1()
res <- Fig9.2(out)
head(sort(res$Stability[,20], decreasing=TRUE))
}
