% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-toy.R
\name{Fig9.3}
\alias{Fig9.3}
\title{Reproduce Figure 9.3}
\usage{
Fig9.3(N = 100, seed = 1)
}
\arguments{
\item{N}{Number of simulated realizations}

\item{seed}{Random number seed for reproducibility}

\item{out}{Output of Ex9.1}
}
\description{
Reproduces Figure 9.3 from the book.  If you specify any options, your results may look different.
}
\examples{
out <- Ex9.1()
res <- Fig9.3(out)
covered <- out$beta >= res$Lower & out$beta <= res$Upper
mean(covered)
table(out$varType, covered)
}
