% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ortho.r
\name{gen_ortho}
\alias{gen_ortho}
\title{Simulate data for regression models in which the design matrix X is orthonormal}
\usage{
gen_ortho(
  n,
  p,
  p1 = floor(p/2),
  beta,
  family = c("gaussian", "binomial"),
  SNR = 1,
  signal = c("homogeneous", "heterogeneous")
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of features}

\item{p1}{Number of nonzero features}

\item{beta}{Vector of regression coefficients in the generating model, or, if a scalar, the value of each nonzero regression coefficient.}

\item{family}{Generate `y according to linear "gaussian" or logistic "binomial" model}

\item{SNR}{Signal to noise ratio}

\item{signal}{Should the beta coefficients be homogeneous (default) or heterogeneous}
}
\description{
Simulate data for regression models in which the design matrix X is orthonormal
}
\examples{
Data <- gen_ortho(5, 2)
Data$X
crossprod(Data$X)
}
