% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genData}
\alias{genData}
\title{Simulate data for regression models}
\usage{
genData(n, p, p1 = floor(p/2), beta, family = c("gaussian",
  "binomial"), SNR = 1, signal = c("homogeneous", "heterogeneous"),
  corr = c("exchangeable", "autoregressive"), rho = 0)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of features}

\item{p1}{Number of nonzero features}

\item{beta}{Vector of regression coefficients in the generating model, or, if a scalar, the value of each nonzero regression coefficient.}

\item{family}{Generate \code{y} according to linear "gaussian" or logistic "binomial" model}

\item{SNR}{Signal to noise ratio}

\item{signal}{Should the beta coefficients be homogeneous (default) or heterogeneous}

\item{corr}{Correlation structure between features ('exchangeable' | 'autoregressive')}

\item{rho}{Correlation coefficient}
}
\description{
Simulate data for regression models
}
\examples{
Data <- genData(100, 100, 10)
dim(Data$X)
head(Data$y)
head(Data$beta)

genData(100, 10, 5)$beta
genData(100, 10, 5, SNR=2)$beta
genData(100, 10, 5, SNR=2, corr='exch', rho=0.7)$beta
genData(100, 10, 5, SNR=2, corr='auto', rho=0.7)$beta
genData(100, 10, 5, SNR=2, corr='auto', rho=0.7, signal='het')$beta
genData(100, 10, 5, SNR=2, corr='auto', rho=0.1, signal='het')$beta
genData(100, 10, 5, SNR=2, corr='auto', rho=0.1, signal='het', b=1)$beta

genData(10, 10, 5, family='binomial')$y

\dontshow{\dontrun{
  require(magrittr)
  genData(1000, 10, rho=0.0, corr='exch')$X \%>\% cor \%>\% round(digits=2)
  genData(1000, 10, rho=0.7, corr='exch')$X \%>\% cor \%>\% round(digits=2)
  genData(1000, 10, rho=0.7, corr='auto')$X \%>\% cor \%>\% round(digits=2)

  genData(1000, 3, 3, rho=0)$X \%>\% cor \%>\% round(digits=2)
}}
}
