% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDataABN.R
\name{genDataABN}
\alias{genDataABN}
\title{Simulate data according to Causal/Correlated/Noise paradigm}
\usage{
genDataABN(n = 100, p = 60, a = 6, b = 2, rho = 0.5,
  family = c("gaussian", "binomial"), signal = c("homogeneous",
  "heterogeneous"), noise = c("exchangeable", "autoregressive"),
  rho.noise = 0, beta, SNR = 1)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of groups}

\item{a}{Number of causal ('A') variables}

\item{b}{Number of correlated ('B') variables}

\item{rho}{Correlation between 'A' and 'B' variables}

\item{family}{Generate \code{y} according to linear "gaussian" or logistic "binomial" model}

\item{signal}{Should the groups be heterogeneous (in beta) or homogeneous?}

\item{noise}{Correlation structure between features ('exchangeable' | 'autoregressive')}

\item{rho.noise}{Correlation parameter for noise variables}

\item{beta}{Vector of regression coefficients in the generating model, or, if a scalar, the value of each nonzero regression coefficient}

\item{SNR}{Signal to noise ratio}
}
\description{
Simulate data according to Causal/Correlated/Noise paradigm
}
\examples{
Data <- genDataABN(n=100, p=20, a=2, b=3)
dim(Data$X)
length(Data$y)
with(Data, data.frame(beta, varType))

genDataABN(100, 10, 2, 1)$beta
genDataABN(100, 10, 2, 1, rho=0.9)$beta
genDataABN(100, 10, 2, 1, rho=0.9, rho.noise=0.9)$beta
genDataABN(100, 10, 2, 1, SNR=3)$beta
genDataABN(100, 10, 2, 1, SNR=3, signal='het')$beta
genDataABN(100, 10, 2, 1, beta=3)$beta
genDataABN(100, 10, 2, 1, beta=10:1)$beta

genDataABN(10, 20, 2, 3, family='binomial')$y

\dontshow{\dontrun{
  require(magrittr)
  genDataABN(1000, 10, 2, 2, rho=0.5, rho.noise=0.0, noise='exch')$X \%>\% cor \%>\% round(digits=2)
  genDataABN(1000, 10, 2, 2, rho=0.5, rho.noise=0.5, noise='exch')$X \%>\% cor \%>\% round(digits=2)
  genDataABN(1000, 10, 2, 2, rho=0.5, rho.noise=0.9, noise='auto')$X \%>\% cor \%>\% round(digits=2)
}}
}
