% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDataGrp.R
\name{genDataGrp}
\alias{genDataGrp}
\title{Simulate grouped data for regression models}
\usage{
genDataGrp(n, J, K = 1, beta, family = c("gaussian", "binomial"),
  J1 = ceiling(J/2), K1 = K, SNR = 1, signal = c("homogeneous",
  "heterogeneous"), signal.g = c("homogeneous", "heterogeneous"), rho = 0,
  rho.g = rho)
}
\arguments{
\item{n}{Sample size}

\item{J}{Number of groups}

\item{K}{Number of features per group}

\item{beta}{Vector of regression coefficients in the generating model, or, if a scalar, the value of each nonzero regression coefficient}

\item{family}{Generate \code{y} according to linear "gaussian" or logistic "binomial" model}

\item{J1}{Number of nonzero groups}

\item{K1}{Number of nonzero coefficients per group}

\item{SNR}{Signal to noise ratio}

\item{signal}{Should the groups be heterogeneous (in beta) or homogeneous?}

\item{signal.g}{Should the coefficients within a group be heterogeneous or homogeneous?}

\item{rho}{Correlation between groups}

\item{rho.g}{Correlation between parameters within a group}
}
\description{
Simulate grouped data for regression models
}
\examples{
Data <- genDataGrp(100, 10, 5, J1=3, K1=2)
dim(Data$X)
head(Data$y)
Data$beta
Data$group

genDataGrp(100, 3, 3, J1=2, K1=2)$beta
genDataGrp(100, 3, 3, J1=2, K1=2, SNR=2)$beta
genDataGrp(100, 3, 3, J1=2, K1=2, SNR=2, rho=0.8)$beta
genDataGrp(100, 3, 3, J1=2, K1=2, SNR=2, rho=0.8, signal='het')$beta
genDataGrp(100, 3, 3, J1=2, K1=2, SNR=2, rho=0.8, signal='het', signal.g='het')$beta
genDataGrp(100, 3, 3, J1=2, K1=2, SNR=2, rho=0.8, signal='het', b=1)$beta

\dontshow{\dontrun{
  require(magrittr)
  genDataGrp(1000, 3, 3, rho=0)$X \%>\% cor \%>\% round(digits=2)
  genDataGrp(1000, 3, 3, rho=0.7)$X \%>\% cor \%>\% round(digits=2)
  genDataGrp(1000, 3, 3, rho=0.3, rho.g=0.8)$X \%>\% cor \%>\% round(digits=2)
}}
}
