% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_utils.R
\name{get_hypothesis_mult}
\alias{get_hypothesis_mult}
\title{Build the hypothesis matrix for a multifactorial setting}
\usage{
get_hypothesis_mult(hypothesis, a, d)
}
\arguments{
\item{hypothesis}{\itemize{
\item either a character or a list with components TW and TS
}}

\item{a}{\itemize{
\item natural number, representing the number of groups
}}

\item{d}{\itemize{
\item natural number, representing the dimension of the observation vectors
}}
}
\value{
Returns a list  with the components

\item{TW}{wholeplot hypothesis matrix TW}

\item{TS}{subplot hypothesis matrix TS}
}
\description{
Since there are two internal functions, for better clarity we create a function
that extracts the hypothesis matrices. This way, changes don't need to be made
in both internal functions.
}
\keyword{internal}
