% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_group_case_generics.R
\name{hdrm1.data.frame}
\alias{hdrm1.data.frame}
\title{Test for one group high dimensional repeated measures}
\usage{
\method{hdrm1}{data.frame}(data, hypothesis = c("flat", "equal"), value, subject, factor, ...)
}
\arguments{
\item{data}{either a matrix with subjects in columns and factor levels in rows or a data.frame of longtable format}

\item{hypothesis}{a character or a quadratic matrix}

\item{value}{if \code{data} is a \code{data.frame}: name or number of value column.}

\item{subject}{if \code{data} is a \code{data.frame}: name or number of subject column.}

\item{factor}{if \code{data} is a \code{data.frame}: name or number of factor column.}

\item{...}{further arguments are currently ignored.}
}
\value{
Returns a list with class "hdrm1" with the components

\item{data}{the initial input data. If a data.frame was given it will have been transformed to a matrix.}

\item{f}{the degrees of freedom \eqn{f} of the distribution of the test statistic}

\item{statistic}{the test statistic \eqn{W}}

\item{tau}{the convergence parameter \eqn{\tau}}

\item{H0}{the Nullhypothesis tested. Will be 'custom' if \code{hypothesis} was given as a matrix.}

\item{H}{the hypothesis-matrix used.}

\item{p.value}{the p-value of the test statistic}

\item{critical.value}{the critical value depending on \eqn{\alpha}}

\item{dim}{a vector of length 2, giving the dimensions \eqn{d \times N} of the data}

\item{removed.cases}{number of subjects removed for having missing values.}

\insertNoCite{Pauly2015}{hdrm}
}
\description{
S3 generic methods for performing a one sample test for one group repeated measure data
}
\details{
If \code{data} is a matrix or a list of multiple matrices with the same amount of rows then it is assumed that individuals are represented by columns and factor levels by rows.

If \code{data} is a data.frame, \code{formula} must be a formula object of the form \code{value ~ subject + factor} that specifies the corresponding columns of the \code{data.frame}.
Each subject and factor level must be represented by a unique ID.

`hypothesis`` can either be given as one of "flat" or "equal" for testing whether the time profile is flat: \eqn{H_0: \mu_1 = \ldots = \mu_d} or equal: \eqn{H_0: \mu = 0}. Alternatively 'hypothesis' can also be a custom contrast matrix \eqn{T \in \mathbb{R}^{d\times d}}

Subjects with missing values will be disregarded.
}
\references{
\insertAllCited
}
