% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_group_case.R
\name{hdrm1_internal}
\alias{hdrm1_internal}
\title{interne Funktion für die Berechnung}
\usage{
hdrm1_internal(X, hypothesis, alpha, na.action = "na.omit")
}
\description{
interne Funktion für die Berechnung
}
\keyword{internal}
