% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_grouped_generics.R
\name{hdrm_grouped.data.frame}
\alias{hdrm_grouped.data.frame}
\title{Test for multiple group high dimensional repeated measures}
\usage{
\method{hdrm_grouped}{data.frame}(
  data,
  hypothesis = c("whole", "sub", "interaction"),
  group,
  value,
  subject,
  factor,
  bootstrap = FALSE,
  B = "500*N",
  ...
)
}
\arguments{
\item{data}{a data.frame in longtable format.}

\item{hypothesis}{either one of "whole", "sub" and "interaction" or a named list with elements TW and TS.}

\item{group}{name or number of corresponding column.}

\item{value}{name or number of value column.}

\item{subject}{name or number of subject column.}

\item{factor}{name or number of factor column.}

\item{B}{a string specifying a function of the number of subjects \eqn{N}. Determinates how many subsamples are used for estimating C5.}

\item{...}{further arguments. Currently ignored}
}
\value{
a named list of class "hdrm_test" with the components

\item{data}{the initial input data converted to a matrix.}

\item{f}{the degrees of freedom \eqn{f} of the critical value.}

\item{statistic}{the test statistic \eqn{W} depending on \eqn{\alpha}.}

\item{tau}{the convergence parameter \eqn{\tau}}

\item{p.value}{the \eqn{p}-value of the test statistic}

\item{dim}{a named list with components d and N of the input data}

\item{groups}{a named list with components a and table giving the number and distribution of groups}

\item{removed.cases}{number of incomplete subjects removed.}

\insertNoCite{Sattler2018}{hdrm}
}
\description{
This is the longtable version. For widetable version see \link{hdrm_grouped.data.frame}.
}
\details{
\code{hypothesis} can either be given as a character or a list. Legal characters are "whole" (default), "sub" or "interaction". "whole" and "sub" stand for \eqn{H_0}: no differences in wholeplot/subplot-factor, while "interaction" means \eqn{H_0}: no interaction between subplot- and wholeplot-factor.
Alternatively \code{hypothesis} can be a named list with components \code{TW} for the wholeplot-part and \code{TS} for the subplot-part of \eqn{T = T_W \otimes T_S}.

Note, that the \code{p.value} may depend heavily on the choice of \code{B}, as both depend on the degrees of freedom \code{f}, which are estimated by C5.
}
\references{
\insertAllCited
}
