% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_grouped.R
\name{hdrm_grouped_eq_cov_internal}
\alias{hdrm_grouped_eq_cov_internal}
\title{Internal function to conduct the grouped test for equal covariances}
\usage{
hdrm_grouped_eq_cov_internal(
  data,
  group,
  hypothesis = c("whole", "sub", "interaction"),
  AM,
  B,
  subsampling,
  seed
)
}
\arguments{
\item{data}{the data for which the test is applied, where two formats are
possible:
-A matrix where subjects represented by columns.
-A value column from a data frame, where in this case an additional parameter
specifying the subjects is necessary.}

\item{group}{vector, which gives the allocation of the matrix cols
to the single groups}

\item{hypothesis}{either one of "whole", "sub" "interaction", "identical" or
"flat" or a named list with quadratic matrices \code{TW} and \code{TS} (see details).}

\item{AM}{binary variable, specifying whether an alternative hypothesis
matrix based on
Sattler and Rosenbaum (2025) should be used, which has
fewer rows but does not influence any values. The usage of this matrix is
the predefined setting.}

\item{B}{\code{integer} that determines the number of subsamples used by the
subsampling trace estimators.}

\item{subsampling}{\code{logical}. Specifying whether the subsampling versions for
all trace estimators should be used (see details).}

\item{seed}{an optional natural number as seed, if it should be set for
reproducibility.}
}
\value{
a named list of class "hdrm_grouped" with the components

\item{data}{the input data used.}

\item{f}{the degrees of freedom \eqn{f}.}

\item{tau}{the convergence parameter \eqn{\tau}.}

\item{H}{a named list with components \code{TW} and \code{TS} that give the
components of the hypothesis matrix.}

\item{hypothesis}{a character. Will be "custom" if \code{hypothesis} is a
list, otherwise \code{hypothesis[1]}.}

\item{p.value}{the \eqn{p}-value of the test statistic.}

\item{dim}{a named list with with number of factor levels \eqn{d} and
number of subjects \eqn{N} of \code{data}.}

\item{groups}{a named list with components number of groups \code{a} and
distribution of groups \code{table}.}

\item{removed.cases}{number of incomplete subjects removed.}

\item{subsamples}{number of subsamples used for subsampling
estimators.}
}
\description{
If all parameter and options are suitable  when executing the function
hdrm_grouped and after preparing the arguments, this function conducts the test,
if equal covariance matrices are assumed.
}
\keyword{internal}
