% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_single_methods.R
\name{hdrm_single_data.frame}
\alias{hdrm_single_data.frame}
\title{Test for one group high dimensional repeated measures}
\usage{
hdrm_single_data.frame(
  data,
  hypothesis = c("flat", "equal"),
  value,
  subject,
  factor,
  ...
)
}
\arguments{
\item{data}{a data.frame in longtable fromat.}

\item{hypothesis}{either a character "flat", "equal" or a quadratic numeric matrix.}

\item{value}{name or number of value column.}

\item{subject}{name or number of subject column.}

\item{factor}{name or number of factor column.}

\item{...}{further arguments are currently ignored.}
}
\value{
Returns a list with class "hdrm_single" with the components

\item{data}{the initial input data.}

\item{f}{the degrees of freedom \eqn{f}.}

\item{statistic}{the test statistic \eqn{W}.}

\item{tau}{the convergence parameter \eqn{\tau}.}

\item{H}{the hypothesis-matrix used.}

\item{hypothesis}{a character. Will be 'custom' if \code{hypothesis} was
given as a matrix, otherwise \code{hypothesis[1]}.}

\item{p.value}{the p-value of the test statistic}

\item{dim}{a named vector giving the dimensions \eqn{d \times N} of \code{data}.}

\item{removed.cases}{number of subjects removed for having missing values.}
}
\description{
This function implements the methods outlined in
\insertCite{Pauly2015;textual}{hdrm} for data in a longtable format. For data in
a widetable format see \link{hdrm_single_matrix}.
}
\details{
The function can deal with missing values only in the \code{value} column.
Affected subjects are dropped with a warning. Missing values in any other
column of data will result in an error.

The test outlined in \insertCite{Pauly2015;textual}{hdrm} is performed for
the hypothesis given by \code{hypothesis}. The \code{hypothesis} can either be given as
a \code{character} or a \code{list}. Legal characters are "flat" (default) and "equal"
"flat" and "equal" stand for \eqn{H_0}: flat(equal time profile. Other
characters will result in an error.

Alternatively \code{hypothesis} can be a quadratic hypothesis matrix \eqn{T} with
the number of rows equal to the number of rows of \code{data}. \eqn{T} must be
idempotent, meaning symmetrical and \eqn{T^2 = T}. A matrix that does not
match those criteria will result in an error.
}
\references{
\insertAllCited
}
