% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_single_methods.R
\name{hdrm_single_widetable}
\alias{hdrm_single_widetable}
\title{Test for one group high dimensional repeated measures}
\usage{
hdrm_single_widetable(data, hypothesis = "flat", ...)
}
\arguments{
\item{data}{a data.frame with subjects represented by columns and factor levels
represented by rows.}

\item{hypothesis}{either "equal" or a quadratic numeric matrix.}

\item{...}{further arguments are currently ignored.}
}
\value{
Returns a list with class "hdrm_single" with the components

\item{data}{the initial input data.}

\item{f}{the degrees of freedom \eqn{f}.}

\item{statistic}{the test statistic \eqn{W}.}

\item{tau}{the convergence parameter \eqn{\tau}.}

\item{H}{the hypothesis-matrix used.}

\item{hypothesis}{a character. Will be 'custom' if \code{hypothesis} was
given as a matrix, otherwise \code{hypothesis[1]}.}

\item{p.value}{the p-value of the test statistic.}

\item{dim}{a named vector giving the dimensions \eqn{d \times N} of \code{data}.}

\item{removed.cases}{number of subjects removed for having missing values.}
}
\description{
This function implements the methods outlined in
\insertCite{Pauly2015;textual}{hdrm} for data in a widetable format. For
data in a longtable format see \link{hdrm_single_longtable}.
}
\details{
If \code{data} contains missing values, affected subjects will be dropped with a
warning.

The test outlined in \insertCite{Pauly2015;textual}{hdrm} is performed for
the hypothesis given by \code{hypothesis}. The \code{hypothesis} can either be given as
a \code{character} or a \code{list}. The only legal character is "flat" (default).
"flat" stands for \eqn{H_0}: time profile is flat. Other
characters will result in an error.

Alternatively, \code{hypothesis} can be the quadratic hypothesis matrix \eqn{T} with
the number of rows equal to the number of rows of \code{data}. \eqn{T} must be
a projection matrix, meaning symmetrical and \eqn{T^2 = T}. A matrix that does not
match those criteria will result in an error.
}
\examples{
data(birthrates)

# call test
hdrm_single_widetable(data = birthrates,
                       # test if time profiles are flat
                       hypothesis = "flat"
)


# hypothesis as list
hypothesis_matrix <- diag(34) - matrix(1/34, 34, 34)

hdrm_single_widetable(data = birthrates,
                      # equivalent to hypothesis = "flat"
                      hypothesis = "flat"
)

rm(hypothesis_matrix)
}
\references{
\insertAllCited
}
