% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_group_case.R
\name{hdrm_test}
\alias{hdrm_test}
\title{dings}
\usage{
hdrm_test(
  data,
  hypothesis = c("whole", "sub", "interaction"),
  alpha = 0.05,
  B = "500*N",
  subsample = FALSE,
  ...
)
}
\description{
dings
}
