% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_group_case.R
\name{hdrm_test_internal}
\alias{hdrm_test_internal}
\title{alibi dokumentation}
\usage{
hdrm_test_internal(
  data,
  group,
  hypothesis = c("whole", "sub", "interaction"),
  alpha = 0.05,
  B = "500*N",
  subsample = subsample
)
}
\description{
alibi dokumentation
}
