% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_axis}
\alias{log_axis}
\title{Make a log-scale axis}
\usage{
log_axis(side = 1, base, disp = base, n = 5, style = NULL, ...)
}
\arguments{
\item{side}{1=bottom, 2=left, 3=top, 4=right}

\item{base}{What base is the log?  Default is base e (natural log).}

\item{disp}{By default, displays in same base as how constructed, but can display in other bases too}

\item{n}{Number of tick marks}

\item{style}{Either "dec" for decimals such as 0.125, or "pow" for powers such as 2^(-3)}

\item{...}{Further arguments to \code{axis()}}
}
\description{
Make a log-scale axis
}
\examples{
x <- rexp(100)
y <- runif(100)
plot(log(x), y)
plot(log(x), y, xaxt='n', xlab='x')
log_axis()
plot(log(x), y, xaxt='n', xlab='x')
log_axis(style='pow', disp=2)
}
