% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_utils.R
\name{MSrootcompact}
\alias{MSrootcompact}
\title{Compact root of a matrix}
\usage{
MSrootcompact(H)
}
\arguments{
\item{H}{a matrix which compact root should be determined}
}
\value{
a compact root of the original matrix
}
\description{
calculates the compact root of a matrix, as it was introduced in
Sattler and Rosenbaum (2025).
}
\keyword{internal}
