% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pal}
\alias{pal}
\title{Generate a nice palette of colors}
\usage{
pal(n, alpha = 1)
}
\arguments{
\item{n}{Number of colors to generate, evenly spaced on hcl scale}

\item{alpha}{Partial transparency}
}
\description{
Generate a nice palette of colors
}
\examples{
pal(3)
}
