% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{plot.ridge}
\alias{plot.ridge}
\title{Plot a ridge regression coefficient path}
\usage{
\method{plot}{ridge}(x, xaxis = c("loglam", "df", "both"), xlab, ylab, ...)
}
\arguments{
\item{x}{a ridge}
}
\description{
Plot a ridge regression coefficient path
}
