% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.r
\name{run}
\alias{run}
\title{Run an example and save the result, or load if already run}
\usage{
run(expr, path)
}
\arguments{
\item{expr}{An R expression that returns a \code{data.frame}-compatible object;
see \verb{[arrow::write_parquet()]}}

\item{path}{A file path (where to save/load the object)}
}
\description{
To trigger a re-run, delete the file. All results are saved as parquet files.
}
\examples{
run(Ex1.1(N=10), '~/a.pqt')
}
