% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seedlings.R
\docType{data}
\name{Seedlings}
\alias{Seedlings}
\title{Seedlings}
\format{
An object of class \code{data.frame} with 48 rows and 12 columns.
}
\source{
Nemec, A F. Analysis of repeated measures and times series: An introduction with forestry examples. Biometrics information handbook No. 6, and Working paper No. 15/1996. Canada.
}
\usage{
Seedlings
}
\description{
Repeated Measures of
}
\keyword{datasets}
