% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunspots_matrix.R
\docType{data}
\name{sunspots_matrix}
\alias{sunspots_matrix}
\title{Monthly Sunspot Numbers, 1749-1983}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 235 rows and 12 columns.
}
\usage{
sunspots_matrix
}
\description{
Monthly mean relative sunspot numbers from 1749 to 1983 \insertCite{sunspots_dataset}{hdrm}. Collected at Swiss Federal Observatory, Zurich until 1960, then Tokyo Astronomical Observatory.
The object \code{sunspots_matrix} is the object \link{sunspots} from the package \link{datasets} converted to a matrix, to make it readily available for the methods in this package.
}
\references{
\insertAllCited
}
\keyword{datasets}
