% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-summary.R
\name{Tab1.1}
\alias{Tab1.1}
\title{Reproduce Table 1.1}
\usage{
Tab1.1()
}
\description{
Reproduces Table 1.1 from the book.  If you specify any options, your results may look different.
}
\examples{
Tab1.1()
}
