% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-tcga-enet.r
\name{Tab4.1}
\alias{Tab4.1}
\title{Reproduce Table 4.1}
\usage{
Tab4.1(seed = 1)
}
\arguments{
\item{seed}{For reproducibility}
}
\description{
Reproduces Table 4.1 from the book.  If you specify any options, your results may look different.
}
\examples{
Tab4.1()

}
