% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bardet.R
\name{Tab4.2}
\alias{Tab4.2}
\title{Reproduce Table 4.2}
\usage{
Tab4.2(p = 5000, seed = 12)
}
\arguments{
\item{p}{Restrict analysis to the `p` genes with the largest variances in expression}

\item{seed}{For reproducibility}
}
\description{
Reproduces Table 4.2 from the book.  If you specify any options, your results may look different.
}
\examples{
Tab4.2()

}
