% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{toplegend}
\alias{toplegend}
\alias{rightlegend}
\title{Make a legend on top or right side of plot}
\usage{
toplegend(horiz = TRUE, ...)

rightlegend(...)
}
\arguments{
\item{horiz}{Horizontal layout?}

\item{...}{Further arguments to \code{legend()}}
}
\description{
Make a legend on top or right side of plot
}
\examples{
g <- rbinom(20, 1, 0.5)
x <- rnorm(20)
y <- rnorm(20)
plot(x, y, col=pal(2)[g+1], pch=19)
toplegend(legend=c('Group 1', 'Group 2'), col=pal(2), pch=19)
}
