% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoari.r
\name{whoari}
\alias{whoari}
\title{World Health Organization study on acute respiratory illnesses}
\format{
\itemize{
\item \code{y}: A pneumonia score, measured on the following scale: 1=No disease,
2=Cold/cough, 3=Pneumonia, 4=Severe pneumonia, 5=Life-threatening illness.
\item \code{X}: A matrix with 816 rows and 66 columns. Each row of \code{X} represents an
individual. The columns are described below.
\itemize{
\item The following columns are objective clinical measurements:
\itemize{
\item \code{biwt}: Birth weight
\item \code{hcir}: Head circumference
\item \code{wght}: Weight (in grams)
\item \code{lgth}: Length (in centimeters)
\item \code{temp}: Temperature (in Celsius)
\item \code{hrat}: Heart rate
\item \code{age}: Age (in days)
\item \code{rr}: Adjusted respiratory rate
\item \code{waz}: Weight-for-age (z-score)
}
\item There are also a number of clinical observations, judged on a severity scale
with higher numbers denoting increased severity. For example, 0-3 would
represent absent/mild/moderate/severe:
\itemize{
\item \code{omph}: omphalitis (0-1)
\item \code{conj}: conjunctivitis (0-1)
\item \code{slpm}: sleeping more (0-1)
\item \code{slpl}: sleeping less (0-1)
\item \code{wake}: wakes less easy (0-1)
\item \code{convul}: history of convulsion (0-1)
\item \code{hfa}: history of abnormal feeding (0-3)
\item \code{hfb}: fast breathing (0-1)
\item \code{hfe}: history of fever (0-1)
\item \code{hap}: history of stop breathing (0-1)
\item \code{hcl}: less activity (0-1)
\item \code{hcm}: crying more (0-1)
\item \code{hcs}: crying less (0-1)
\item \code{hdi}: history of diarrhoea (0-1)
\item \code{hvo}: vomit more (0-1)
\item \code{bat}: birth preterm (0-1)
\item \code{fde}: fever at delivery (0-1)
\item \code{chi}: previous child died (0-1)
\item \code{hbr}: engorged breasts (0-1)
\item \code{twb}: water broke (0-1)
\item \code{ldy}: days in labour (1-9)
\item \code{inc}: respiratory distress (0-2)
\item \code{sr1}: respiratory state (0-3)
\item \code{sr2}: respiratory state (0-3); I do not know why this is measured twice
\item \code{apn}: apnea (0-1)
\item \code{lcw}: lower chest in-drawing (0-3)
\item \code{nfl}: nasal flaring (0-3)
\item \code{str}: stridor (0-2)
\item \code{gru}: grunting (0-2)
\item \code{coh}: cough heard (0-1)
\item \code{ccy}: central cyanosis (0-1)
\item \code{jau}: jaundice (0-1)
\item \code{csd}: drowsy state (0-2)
\item \code{csa}: agitated state (0-1)
\item \code{aro}: arousal (0-2)
\item \code{qcr}: quality of crying (0-2)
\item \code{con}: consolability (0-2)
\item \code{att}: attentive (0-2)
\item \code{mvm}: amount of movement (0-2)
\item \code{afe}: drinking ability (0-2)
\item \code{absu}: sucking ability (0-2)
\item \code{stu}: skin turgor (0-2)
\item \code{deh}: dehydrated (0-2)
\item \code{dcp}: digital capillary refill (0-2)
\item \code{crs}: crepitation (0-2)
\item \code{skr}: skin rash (0-2)
\item \code{abb}: bulging fontanelle (0-1)
\item \code{abk}: sunken fontanelle (0-1)
\item \code{hyp}: hypotonia (0-3)
\item \code{whz}: wheezing (0-1)
\item \code{hdb}: difficulty breathing (0-1)
\item \code{smi2}: smiling ability
\item \code{puskin}: pustular skin rash (0-1)
\item \code{abd}: abdominal distension (0-4)
\item \code{nut}: nutritional status score (0-3)
\item \code{oto}: otoscopy impression (0-2)
\item \code{oab}: sclerema (0-1)
}
}
}
}
\source{
I obtained the data from
\url{http://biostat.mc.vanderbilt.edu/wiki/Main/DataSets}. The
original reference is:

Harrell F, Margolis P, Gove S, Mason K, Mulholland E, Lehmann D, Muhe L,
Gatchalian S and Eichenwald H (1998). Development of a clinical prediction
model for an ordinal outcome: the World Health Organization multicentre study
of clinical signs and etiological agents of pneumonia, sepsis and meningitis
in young infants. \emph{Statistics in Medicine}, \strong{17}: 909-944.
}
\description{
The presentation of an acutely ill young infant presents health workers,
especially those in developing countries, with a very difficult problem. Serious
infections are the main cause of morbidity and mortality in infants under 3
months of age in these countries, and diagnosing the severity of the illness is
rather difficult.

To study this problem, the World Health Organization (WHO) collected data on a
number of readily accessible variables such as vital signs, family history, and
clinical observations resulting from physical examination. The patients' disease
status was later determined based on the course of the disease and various
laboratory tests. The goal of the study was to develop a early prediction rule
for grading the severity of the disease so that timely treatment could be
delivered (and costly but unnecessary treatments avoided).

The WHO study looked at several acute respiratory illnesses in several
countries. This data set contains data on pneumonia from the country Ethiopia.
}
