% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spurenschaetzer_Pauly.R
\name{B0}
\alias{B0}
\title{trace estimator B_0}
\usage{
B0(X, N)
}
\arguments{
\item{X}{data matrix with individuals in rows and levels in cols multiplied with the hypothesis matrix}

\item{N}{number of individuals}
}
\value{
returns the trace estimator for T*V
}
\description{
trace estimator B_0
}
\keyword{internal}
