% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spurenschaetzer_Pauly.R
\name{B2}
\alias{B2}
\title{trace estimator B_2}
\usage{
B2(X, N)
}
\arguments{
\item{X}{data matrix with individuals in rows and levels in cols multiplied with the hypothesis matrix}

\item{N}{number of individuals}
}
\value{
returns the trace estimator for (T*V)^2
}
\description{
trace estimator B_2
}
\keyword{internal}
