% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spurenschaetzer_Pauly.R
\name{B3}
\alias{B3}
\title{trace estimator B_3}
\usage{
B3(X)
}
\arguments{
\item{X}{data matrix with individuals in rows and levels in cols multiplied with the hypothesis matrix}
}
\value{
returns the trace estimator for (T*V)^3
}
\description{
trace estimator B_3
}
\keyword{internal}
