% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthrates.R
\docType{data}
\name{birthrates}
\alias{birthrates}
\title{Yearly birthrates in German States 1990-2023}
\format{
An object of class \code{data.frame} with 34 rows and 16 columns.
}
\usage{
birthrates
}
\description{
A dataset containing the number of births per woman in each German State for 1990-2023.
}
\references{
Statistisches Bundesamt (Destatis), Genesis-Online, \href{https://www-genesis.destatis.de/genesis//online?operation=table&code=12612-0012&bypass=true&levelindex=0&levelid=1726659625856#abreadcrumb}{Table 12612-0009}. Last visit 18.09.2024; \href{www.govdata.de/dl-de/by-2-0}{Datenlizenz by-2-0}.
}
\keyword{datasets}
