% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthrates.R
\docType{data}
\name{birthrates}
\alias{birthrates}
\title{Yearly birthrates in German States 1990-2023}
\format{
An object of class \code{data.frame} with 34 rows and 16 columns.
}
\usage{
birthrates
}
\description{
A dataset containing the number of births per woman in each German State for 1990-2023.
}
\references{
Statistisches Bundesamt (Destatis). (2023). Statistischer Bericht - Geburten, Tabelle 12612-09. Statisitsches Bundesamt.
}
\keyword{datasets}
