% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthrates.R
\docType{data}
\name{birthrates}
\alias{birthrates}
\title{Yearly birthrates in German States 1990-2023}
\format{
An object of class \code{data.frame} with 34 rows and 16 columns.
}
\usage{
birthrates
}
\description{
A data set containing the number of births per woman in each German State for
1990-2023 \insertCite{birthrates}{hdrm}. States are represented by columns
and time is represented by rows.
}
\references{
{\insertAllCited{}}
}
\keyword{datasets}
