% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthrates.R
\docType{data}
\name{birthrates}
\alias{birthrates}
\title{Yearly birthrates in German States 1990-2023}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 34 rows and 16 columns.
}
\usage{
birthrates
}
\description{
A data set containing the number of births per woman in each German State for
1990-2023 (Statistisches Bundesamt (Destatis), 2024). States are represented by columns
and time is represented by rows.
}
\references{
{Statistisches Bundesamt (Destatis) (2024). “Statistischer Bericht - Geburten 2023, Tabelle 12612-09.” https://www.destatis.de/DE/Themen/Gesellschaft-Umwelt/Bevoelkerung/Geburten/Publikationen/Downloads-Geburten/statistischer-bericht-geburten-5126104237005.html.}
}
\keyword{datasets}
