% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-glmnet.r
\name{boot.glmnet}
\alias{boot.glmnet}
\title{Calculate confidence intervals for lasso using bootstrap}
\usage{
boot.glmnet(X, y, B = 500, lambda, seed, alpha = 0.05, bar = TRUE)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response vector}

\item{B}{Number of bootstrap resamples (default 500)}

\item{lambda}{Regularization parameter at which solutions are to be bootstrapped (by default, uses cross-validation to find lambda)}

\item{seed}{Seed (optional)}

\item{alpha}{Error rate; 0.05 (default) corresponds to 95\% confidence interval}

\item{bar}{Print a progress bar?}
}
\description{
Calculate confidence intervals for lasso using bootstrap
}
\examples{
Data <- gen_data_abn(n=100, p=20, a=2, b=3)
CI <- boot.glmnet(Data$X, Data$y)
covered <- Data$beta >= CI[,1] & Data$beta <= CI[,2]
table(covered)
CI <- boot.glmnet(Data$X, Data$y, lambda=0.65)
covered <- Data$beta >= CI[,1] & Data$beta <= CI[,2]
table(covered)
}
