% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIplot.R
\name{CIplot}
\alias{CIplot}
\alias{CIplot.matrix}
\title{Plot confidence intervals (NEEDS DOCUMENTATION)}
\usage{
CIplot(obj, ...)

\method{CIplot}{matrix}(obj, labels = rownames(B), sort = TRUE, pxlim,
  xlim, ylim, sub, diff = (ncol(B) == 4), null = 0, n.ticks = 6, mar,
  axis = !add, trans, p.label = FALSE, xlab = "", ylab = "",
  add = FALSE, setupOnly = FALSE, lwd = 2,
  replaceUnderscore = TRUE, ...)
}
\description{
Plot confidence intervals (NEEDS DOCUMENTATION)
}
