% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive-lasso.R
\name{cv.adaptive_lasso}
\alias{cv.adaptive_lasso}
\title{Adaptive lasso with cross-validation, using lasso(BIC) as initial estimator}
\usage{
cv.adaptive_lasso(X, y, nfolds = 10, seed)
}
\arguments{
\item{X}{Design matrix, as in `glmnet()`}

\item{y}{Response vector, as in `glmnet()`}

\item{nfolds}{Number of cv folds (default: 10)}

\item{seed}{For reproducibility with respect to the fold assignments}
}
\description{
Using `cv.glmnet()` to carry out the adaptive lasso will result in an underestimation of error unless you also apply cross-validation to the initial estimator; this function provides a wrapper to do that.
}
\examples{
attachData(pollution)
cvfit <- cv.adaptive_lasso(X, y)
plot(cvfit)
summary(cvfit)

}
