% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEG.R
\docType{data}
\name{EEG}
\alias{EEG}
\title{EEG data}
\format{
An object of class \code{data.frame} with 6400 rows and 7 columns.
}
\source{
somelink.com
}
\usage{
EEG
}
\description{
EEG data from...
}
\keyword{datasets}
