% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-selection.r
\name{Ex1.1}
\alias{Ex1.1}
\title{Reproduce Example 1.1}
\usage{
Ex1.1(N = 100, n = 25, p = 100, bar = TRUE, seed = 1)
}
\arguments{
\item{N}{Number of simulated realizations}

\item{n}{Sample size}

\item{p}{Number of features}

\item{bar}{Show progress bar?}

\item{seed}{For reproducibility (optional)}
}
\description{
Reproduces Example 1.1 from the book.  If you specify any options, your results may look different.
}
\examples{
Ex1.1(N=10)
}
