% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet-sim-ind.r
\name{Ex4.3}
\alias{Ex4.3}
\alias{Fig4.2}
\title{Reproduce Example 4.3 and Figure 4.2}
\usage{
Ex4.3(N = 100, s = seq(0.1, 1.1, 0.2), n = 100, p = 500, p1 = 12, seed = 1)

Fig4.2(res, parlist = list(mfrow = c(1, 2), mar = c(4, 4, 2, 0.5)))
}
\arguments{
\item{N}{Number of simulated realizations}

\item{s}{Signal strength (coefficient value for the non-null features)}

\item{n}{Sample size}

\item{p}{Number of features}

\item{p1}{Number of non-null features}

\item{seed}{Seed for reproducibility}

\item{res}{Output from Ex4.3}

\item{parlist}{List of arguments to pass to \code{par()}}
}
\description{
Reproduces Example 4.3 and Figure 4.2 from the book. If you specify any options, your results may look different.
}
\examples{
res <- Ex4.3(N=2, s=c(0.3, 1))
Fig4.2(res)
}
