% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot-sim.R
\name{Ex9.3}
\alias{Ex9.3}
\title{Reproduce Example 9.3}
\usage{
Ex9.3(N = 100, B = 100, n = 100, p = 60, n.corr = 2, rho = 0.5,
  seed = 1)
}
\arguments{
\item{N}{Number of simulated realizations}

\item{B}{Number of bootstrap replications (for each data set)}

\item{n}{Sample size}
}
\description{
Reproduces Example 9.3 from the book.  If you specify any options, your results may look different.
}
\examples{
out <- Ex9.3()
}
