% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.r
\name{Fig1.1}
\alias{Fig1.1}
\title{Reproduce Figure 1.1}
\usage{
Fig1.1(n = 20, p = n - 1, seed = 1)
}
\arguments{
\item{n}{Sample size}

\item{p}{Number of features}

\item{seed}{Random seed}
}
\description{
Reproduces Figure 1.1 from the book; if you specify any options, your results may look different.
}
\examples{
Fig1.1()

}
