% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-selection.R
\name{Fig1.2}
\alias{Fig1.2}
\title{Reproduce Figure 1.2}
\usage{
Fig1.2(out)
}
\arguments{
\item{out}{Output of Ex1.1()}
}
\description{
Reproduces Figure 1.2 from the book.  If you specify any options, your results may look different.
}
\examples{
out <- Ex1.1()
Fig1.2(out)
}
