% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge-trace.R
\name{Fig1.4}
\alias{Fig1.4}
\title{Reproduce Figure 1.4}
\usage{
Fig1.4(mar = c(5, 5, 5, 7))
}
\arguments{
\item{mar}{Margins (passed to par)}
}
\description{
Reproduces Figure 1.4 from the book.  If you specify any options, your results may look different.
}
\examples{
Fig1.4()
}
