% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concave-penalties.r
\name{Fig3.1}
\alias{Fig3.1}
\title{Reproduce Figure 3.1}
\usage{
Fig3.1(
  range = c(-4, 4),
  col = c("#FF4E37FF", "#00B500FF", "#008DFFFF"),
  which = c("all", "left", "middle", "right", "all"),
  parlist
)
}
\arguments{
\item{range}{Range for beta coefficient (vector of length 2)}

\item{col}{Lasso/ridge color (vector of length 2)}

\item{which}{\code{left}, \code{middle}, \code{right}, or \code{all}}

\item{parlist}{List of arguments to pass to \code{par()}}
}
\description{
Reproduces Figure 3.1 from the book; if you specify any options, your results may look different.
}
\examples{
Fig3.1()
Fig3.1(which='middle')
}
