% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-tcga-al.R
\name{Fig3.7}
\alias{Fig3.7}
\title{Reproduce Figure 3.7}
\usage{
Fig3.7(seed = 1, ...)
}
\arguments{
\item{seed}{For reproducibility}

\item{...}{Additional arguments to \code{plot.cv.ncvreg()}}
}
\description{
Reproduces Figure 3.7 from the book.  If you specify any options, your results may look different.
}
\examples{
Fig3.7()
}
