% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc-tcga-mcp.r
\name{Fig3.8}
\alias{Fig3.8}
\title{Reproduce Figure 3.8}
\usage{
Fig3.8(seed = 1, nfolds = 10, parlist = list(mfrow = c(2, 2)))
}
\arguments{
\item{seed}{For reproducibility}

\item{nfolds}{As in \code{cv.ncvreg()} (default: 10)}

\item{parlist}{List of arguments to pass to \code{par()}}
}
\description{
Reproduces Figure 3.8 from the book.  If you specify any options, your results may look different.
}
\examples{
Fig3.8(nfolds=3)
}
