% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnet-sim-ind.R
\name{Fig4.2}
\alias{Fig4.2}
\title{Reproduce Figure 4.2}
\usage{
Fig4.2(res, parlist = list(mfrow = c(1, 2), mar = c(4, 4, 2, 0.5)))
}
\arguments{
\item{res}{Output from Ex4.3}

\item{parlist}{List of arguments to pass to \code{par()}}
}
\description{
Reproduces Figure 4.2 from the book.  If you specify any options, your results may look different.
}
\examples{
res <- Ex4.3(N=5, corr='cs')
Fig4.2(res)
}
