% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-split.r
\name{Fig9.1}
\alias{Fig9.1}
\title{Reproduce Figure 9.1}
\usage{
Fig9.1(out, N = 100, seed = 1)
}
\arguments{
\item{out}{Output of Ex9.1()}

\item{N}{Number of random splits}

\item{seed}{Random number seed for reproducibility}
}
\value{
A vector of p-values
}
\description{
Reproduces Figure 9.1 from the book.  If you specify any options, your results may look different.
}
\examples{
Data <- Ex9.1()
out <- Fig9.1(Data)
head(out)
}
