% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm.R
\docType{package}
\name{hdrm-package}
\alias{hdrm}
\alias{hdrm-package}
\title{hdrm: High-Dimensional Regression Modeling}
\description{
Examples, simulations, and figures from the book, "High-Dimensional Regression Modeling", by Breheny and Huang.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{http://github.com/pbreheny/hdrm/issues}
}

}
\keyword{internal}
