% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1dim case.R
\name{hdrm1}
\alias{hdrm1}
\title{Test for one dimensional repeated measures}
\usage{
hdrm1(data, hypothesis = c("equal", "flat"), alpha = 0.05)
}
\arguments{
\item{data}{either a matrix or a data.frame with N subjects in rows and d factor levels in columns}

\item{hypothesis}{either one of c("equal","flat") or a matrix of dimensions c(d,d)}

\item{alpha}{alpha level used for calculating the critical value}
}
\value{
\item{data}{the initial input data}

\item{f}{the degrees of freedom f of the Distribution of the test statistic}

\item{statistic}{the value of the test statistic}

\item{tau}{the convergence parameter tau}

\item{hypothesi}{the type of hypothesis}

\item{critical.value}{the critical value depending on alpha}

\item{p.value}{the $p$-value of the test statistic}

\item{dim}{a vector of length 2, giving the dimensions c(d,N) of the data}
}
\description{
S3 generic methods for performing a one sample test for high dimensional repeated measures
}
