% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_group_case_generics.R
\name{hdrm1}
\alias{hdrm1}
\alias{hdrm1.matrix}
\alias{hdrm1.data.frame}
\title{Test for one group high dimensional repeated measures}
\usage{
hdrm1.matrix(data, hypothesis,...)

hdrm1.data.frame(data, hypothesis, value, subject, factor...)
}
\arguments{
\item{data}{either a matrix with subjects in columns and factor levels in rows or a data.frame of longtable format}

\item{hypothesis}{One of 'equal' or 'flat' for H0: equal/flat time profile. Can also be a matrix of dimensions \eqn{d\times d}}

\item{...}{further arguments are currently ignored.}

\item{formula}{a formula object that specifies the corresponding cols if data is a data.frame. Must be of the form \code{value ~ subject + factor}}

\item{alpha}{alpha level used for calculating the critical value}
}
\value{
Returns a list with class "hdrm1" with the components

\item{data}{the initial input data. If a data.frame was given it will have been transformed to a matrix.}

\item{f}{the degrees of freedom \eqn{f} of the distribution of the test statistic}

\item{statistic}{the test statistic \eqn{W}}

\item{tau}{the convergence parameter \eqn{\tau}}

\item{H0}{the Nullhypothesis tested. Will be 'custom' if \code{hypothesis} was given as a matrix.}

\item{H}{the hypothesis-matrix used.}

\item{p.value}{the p-value of the test statistic}

\item{critical.value}{the critical value depending on \eqn{\alpha}}

\item{dim}{a vector of length 2, giving the dimensions \eqn{d \times N} of the data}

\item{removed.cases}{number of subjects removed for having missing values.}
}
\description{
S3 generic methods for performing a one sample test for one group repeated measure data
}
\details{
If \code{data} is a matrix or a list of multiple matrices with the same amount of rows then it is assumed that individuals are represented by columns and factor levels by rows.

If \code{data} is a data.frame, \code{formula} must be a formula object of the form \code{value ~ subject + factor} that specifies the corresponding columns of the \code{data.frame}.
Each subject and factor level must be represented by a unique ID.

`hypothesis`` can either be given as one of "flat" or "equal" for testing whether the time profile is flat: \eqn{H_0: \mu_1 = \ldots = \mu_d} or equal: \eqn{H_0: \mu = 0}. Alternatively 'hypothesis' can also be a custom contrast matrix \eqn{T \in \mathbb{R}^{d\times d}}

Subjects with missing values will be disregarded.
}
\examples{
## hdrm1 with matrix
N <- 20
d <- 30
# example data with N = 20 subjects in columns and d = 30 factor levels in rows
M <- matrix(rnorm(d*N), d, N)

# test M for H_0 = mu_1 = mu_2 = mu_3 = 0 time profile
H <- matrix(0,d,d)
H[c(1, d+2, 2*d+3)] <- 1
hdrm1(data = M, hypothesis = H)

## hdrm1 with data.frame in longtable format
head(EEG)
# for 'formula' we are interested in value = value, subject = subject and factor = dimension
formula <- value ~ subject + dimension

# test EEG for equal time profile
hdrm1(data = EEG, formula = formula, hypothesis = "equal")
}
\references{
Pauly, M., Ellenberger, D., & Brunner, E. (2015). Analysis of high-dimensional one group repeated measures designs. Statistics, 49(6), 1243–1261. \url{https://doi.org/10.1080/02331888.2015.1050022}
}
