% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_grouped_generics.R
\name{hdrm_grouped}
\alias{hdrm_grouped}
\title{Test for multiple group high dimensional repeated measures.}
\usage{
hdrm_grouped(
  data,
  hypothesis = c("whole", "sub", "interaction"),
  group,
  value,
  subject,
  factor,
  bootstrap = FALSE,
  B = "500*N",
  ...
)
}
\arguments{
\item{data}{either a matrix in widetable or a data.frame in longtable format}

\item{hypothesis}{either one of "whole", "sub" and "interaction" or a named list with elements TW and TS.}

\item{group}{if \code{data} is a matrix: a factor of length N that specifies the groups. If \code{data} is a \code{data.frame}: name or number of corresponding column.}

\item{value}{if \code{data} is a \code{data.frame}: name or number of value column.}

\item{subject}{if \code{data} is a \code{data.frame}: name or number of subject column.}

\item{factor}{if \code{data} is a \code{data.frame}: name or number of factor column.}

\item{bootstrap}{a logical indicating whether the bootstrap version of the trace estimators should be used.}

\item{B}{a string specifying a function of the number of subjects \eqn{N}. Determinates how many subsamples are used for the bootstrap version of the trace estimators.}

\item{...}{further arguments. Currently ignored}
}
\value{
a named list of class "hdrm_test" with the components

\item{data}{the initial input data converted to a matrix.}

\item{f}{the degrees of freedom \eqn{f} of the critical value.}

\item{statistic}{the test statistic \eqn{W} depending on \eqn{\alpha}.}

\item{tau}{the convergence parameter \eqn{\tau}}

\item{p.value}{the \eqn{p}-value of the test statistic}

\item{dim}{a named list with components d and N of the input data}

\item{groups}{a named list with components a and table giving the number and distribution of groups}

\item{removed.cases}{number of incomplete subjects removed.}
}
\description{
See \link{hdrm_grouped.matrix} and \link{hdrm_grouped.data.frame} for data in  wide- and longtable versions, respectively.
}
\details{
\code{hypothesis} can either be given as a character or a list. Legal characters are "whole" (default), "sub" or "interaction". "whole" and "sub" stand for \eqn{H_0}: no differences in wholeplot/subplot-factor, while "interaction" means \eqn{H_0}: no interaction between subplot- and wholeplot-factor.
Alternatively \code{hypothesis} can be a named list with components \code{TW} for the wholeplot-part and \code{TS} for the subplot-part of \eqn{T = T_W \otimes T_S}.

Note that even for \code{bootstrap = FALSE} the bootstrap version for the trace estimator C5 is used.
Therefore the \code{p.value} and the parameter \code{f} are seed dependent and may vary a lot for small B's.
For \eqn{p}-values near \eqn{alpha} it is advised to run the test again with \code{B = "5000*N"} at least.
}
\references{
Paavo Sattler, Markus Pauly. "Inference for high-dimensional split-plot-designs: A unified approach for small to large numbers of factor levels." Electronic Journal of Statistics, 12(2) 2743-2805 2018. https://doi.org/10.1214/18-EJS1465
}
