% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_grouped_methods.R
\name{hdrm_grouped}
\alias{hdrm_grouped}
\title{Test for multiple group high dimensional repeated measures}
\usage{
hdrm_grouped(
  data,
  hypothesis = c("whole", "sub", "interaction", "identical", "flat"),
  AM = TRUE,
  group,
  subject = NULL,
  cov.equal = FALSE,
  subsampling = FALSE,
  B = "1000*N",
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{the data for which the test is applied, where two formats are
possible:
\itemize{
\item A matrix where subjects are represented by columns
\item A value column from a data frame. In this case an additional parameter
specifying the subjects and their groups is necessary.
}}

\item{hypothesis}{either one of "whole", "sub" "interaction", "identical" or
"flat" or a named list with quadratic matrices \code{TW} and \code{TS} (see details).}

\item{AM}{\code{logical}. If \code{TRUE} (default), an alternative hypothesis
matrix based on Sattler and Rosenbaum (2025) is used, which has
fewer rows but does not influence any values.}

\item{group}{parameter vector, which allocates the measurements to the single
groups (see Details).}

\item{subject}{optional parameter vector of \code{length(data)}, which gives the allocation of the
measurements to the subjects (see Details).}

\item{cov.equal}{\code{logical}. If \code{TRUE} (default = \code{FALSE}), the test for equal group covariances is performed as outlined in Sattler (2021).}

\item{subsampling}{\code{logical}. If \code{TRUE} (default = \code{FALSE}), the subsampling versions for
all trace estimators are used (see details).}

\item{B}{a \code{string} of the pattern "1000*N" specifying a function of the
number of subjects \eqn{N}.
Determines the number of subsamples used by the subsampling trace estimators.}

\item{seed}{an optional natural number as seed, if it should be set for reproducibility. Default
is \code{NULL}, which means no seed is set.}

\item{...}{further arguments. Currently ignored.}
}
\value{
a named list of class "hdrm_grouped" with the components

\item{data}{the input data used.}

\item{f}{the degrees of freedom \eqn{f}.}

\item{tau}{the convergence parameter \eqn{\tau}.}

\item{H}{a named list with components \code{TW} and \code{TS} that give the
components of the hypothesis matrix.}

\item{hypothesis}{a character. Will be 'custom' if \code{hypothesis} is a
list, otherwise \code{hypothesis[1]}.}

\item{p.value}{the \eqn{p}-value of the test statistic.}

\item{dim}{a named list with with number of factor levels \eqn{d} and
number of subjects \eqn{N} of \code{data}.}

\item{groups}{a named list with components number of groups \code{a} and
distribution of groups \code{table}.}

\item{removed.cases}{number of incomplete subjects removed.}

\item{subsamples}{number of subsamples used for subsampling
estimators.}

\item{cov.equal}{\code{logical}. Indicates whether the test for equal covariances was used.}
}
\description{
This function implements the methods outlined in
Sattler and Pauly (2018).
}
\details{
If \code{data} is a matrix, the parameter \code{group} allocates the columns to
the groups. Therefore \code{group} must be of length \code{ncol(data)}. \cr
If \code{data} is a vector, \code{subject} allocates the measurements to the subjects
and \code{group} then allocates the subjects to a group. \code{subject} and \code{group}
then must be of length \code{length(data)}.

The function can deal with missing values only in the measurements.
The test is then performed without the affected subjects. Missing values in
\code{group} or \code{subject} will result in an error.

The test outlined in Sattler and Pauly (2018) is performed for
the hypothesis \eqn{(\bm T_W \otimes \bm T_S)\bm\mu = \bm 0}, with \eqn{\bm
T_W}, \eqn{\bm T_S} given by \code{hypothesis}. The \code{hypothesis} can either be
given as a \code{character} or a \code{list}. Legal characters are "whole" (default),
"sub", "interaction", "identical" and "flat". The first three are given by (as
outlined in Sattler and Pauly (2018)):
\itemize{
\item \eqn{H_0^{\overline W}\text{: } \left(\bm P_a \otimes \frac{1}{d}\bm J_d \right)\bm\mu=\bm 0}
(no main effect in whole plot factor group)
\item  \eqn{H_0^{\overline S}\text{: } \left(\frac{1}{a}\bm J_a \otimes \bm P_d \right)\bm\mu=\bm 0}
(no main effect in sub plot factor dimension)
\item \eqn{H_0^{WS}\text{: } \left(\bm P_a \otimes \bm P_d \right)\bm\mu=\bm 0}
(no interaction effect of whole plot and sub plot factor).
}
Additionally, "identical" and "flat" specify the hypotheses
\itemize{
\item  \eqn{H_0^{W}\text{: } \left(\bm P_a \otimes \bm I_d \right)\bm\mu=\bm 0}
(vectors of expected values are equal between all groups)
\item \eqn{H_0^{S}\text{: } \left(\bm I_a \otimes \bm P_d \right)\bm\mu=\bm 0}
(flat profile in all groups).
}
Other characters will result in an error.

Alternatively, \code{hypothesis} can be a named list with quadratic
matrices \code{TW} for the wholeplot-part and \code{TS} for the subplot-part of \eqn{\bm T
= \bm T_W \otimes \bm T_S}. If \code{hypothesis} is a list, \code{TW} and \code{TS} must be
projection matrices, meaning symmetrical and \eqn{\bm T_W^2 = \bm T_W},
\eqn{\bm T_S^2 = \bm T_S}. Differences up to the tolerance in \code{all.equal} are
ignored, while bigger differences will result in a warning. Also, the number of
rows and columns of \code{TW} must be equal to the number of groups and the number
of rows and columns of \code{TS} must be equal to the number of factor levels.
Lists that do not match those criteria will result in an error.

Note that for \code{subsampling = FALSE} results are still seed dependent,
because the computational heaviest trace estimator is always calculated using
subsamples. Also, depending on the choice of \code{B}, the non-subsampling versions
might not be faster for small data.

That also means, that even for \code{subsampling = FALSE} the \code{p.value} depends
heavily on the choice of \code{B}, as it depends on the degrees of freedom \code{f},
which are always estimated by the subsampling version. The test statistic
\eqn{W} is only seed dependent for \code{subsampling = TRUE}

The number of subsamples \code{B} can also be a numeric value. However, this is not
advised, as \code{B} should be a function of \eqn{N} that goes to \eqn{\infty} for
\eqn{N\to\infty}.
}
\examples{
## load data set EEG (data frame)
data(EEG)

# call test
hdrm_grouped(data = EEG$value,
             # test for no group effect
             hypothesis = "whole",
             group = EEG$group,
             ## if data is a vector, subject has to be specified
             subject = EEG$subject,
             subsampling = FALSE,
             B = "100*N"
)


# test using all subsampling version of the trace estimators
hdrm_grouped(data = EEG$value,
             # test for no time effect
             hypothesis = "sub",
             group = EEG$group,
             subject = EEG$subject,
             subsampling = TRUE,
             # B can also be given as a number
             B = 10000
)


# hypothesis as list
hypothesis_list <- list(TW = matrix(1/4, 4, 4),
                        TS = diag(40) - matrix(1/40, 40, 40)
) # equivalent to hypothesis = "sub"

# call test
hdrm_grouped(data = EEG$value,
             hypothesis = hypothesis_list,
             group = EEG$group,
             subject = EEG$subject,
             subsampling = FALSE,
             B = "100*N"
)
rm(hypothesis_list)


## load data set birthrates (matrix)
data(birthrates)
# ?birthrates

## divide states into (former) east and west with Berlin as east
group <- factor(c(1,1,2,2,1,1,1,2,1,1,1,1,2,2,1,2), labels = c("west", "east"))

## call test for interaction effect
hdrm_grouped(data = birthrates,
             ## test for interaction effect between group and dimension
             hypothesis = "interaction",
             group = group,
             ## if data is a matrix, subject does not have to be specified
             B = "100*N"
)


## test using all subsampling version of the trace estimators
hdrm_grouped(data = birthrates,
             ## test for group effect between group and dimension
             hypothesis = "whole",
             group = group,
             ## use the subsampling version of all trace estimators
             subsampling = TRUE,
             ## B can also be given as a number
             B = 10000
)


## hypothesis as list, equivalent to hypothesis = "sub"
hypothesis_list <- list(TW = matrix(1/2, 2, 2),
                        TS = diag(34) - matrix(1/34, 34, 34)
)

## call test with hypothesis as list
hdrm_grouped(birthrates,
             ## equivalent to hypothesis = "sub"
             hypothesis = hypothesis_list,
             group = group,
             subsampling = TRUE,
             B = "100*N"
)

}
\references{
{
Sattler, P., & Pauly, M. (2018). Inference for high-dimensional split-plot-designs: A unified approach for small to large numbers of factor levels. Electronic Journal of Statistics, 12(2), 2743–2805. doi:10.1214/18-ejs1465
}

{
Sattler, P. (2021). A comprehensive treatment of quadratic-form-based inference in repeated measures designs under diverse asymptotics. Electronic Journal of Statistics, 15(1). doi:10.1214/21-ejs1865
}

{
Sattler, P., & Rosenbaum, M. (2025). Choice of the hypothesis matrix for using the Anova-type-statistic. Statistics & Probability Letters, 219(110356), 110356. doi:10.1016/j.spl.2025.110356
}
}
