% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_grouped_methods.R
\name{hdrm_grouped_longtable}
\alias{hdrm_grouped_longtable}
\title{Test for multiple group high dimensional repeated measures}
\usage{
hdrm_grouped_longtable(
  data,
  hypothesis = c("whole", "sub", "interaction"),
  group,
  value,
  subject,
  dimension,
  subsampling = FALSE,
  B = "500*N",
  ...
)
}
\arguments{
\item{data}{a data.frame in longtable format.}

\item{hypothesis}{either one of "whole", "sub" and "interaction" or a named
list with quadratic matrices \code{TW} and \code{TS}.}

\item{group}{name or number of group column.}

\item{value}{name or number of value column.}

\item{subject}{name or number of subject column.}

\item{dimension}{name or number of dimension column.}

\item{subsampling}{\code{logical}. Specifying whether the subsampling versions for all
trace estimators should be used (see details).}

\item{B}{a \code{string} specifying a function of the number of subjects \eqn{N}.
Determines the number of subsamples used by the  subsampling trace estimators.}

\item{...}{further arguments. Currently ignored}
}
\value{
a named list of class "hdrm_grouped" with the components

\item{data}{the input data used.}

\item{f}{the degrees of freedom \eqn{f}.}

\item{tau}{the convergence parameter \eqn{\tau}}

\item{H}{a named list with components \code{TW} and \code{TS} that give the
components of the hypothesis matrix}

\item{hypothesis}{a character. Will be "custom" if \code{hypothesis} is a
list, otherwise \code{hypothesis[1]}.}

\item{p.value}{the \eqn{p}-value of the test statistic.}

\item{dim}{a named list with with number of factor levels \eqn{d}
and number of subjects \eqn{N} of \code{data}}

\item{groups}{a named list with components number of groups \code{a}
and distribution of groups \code{table}.}

\item{removed.cases}{number of incomplete subjects removed.}

\item{subsamples}{number of subsamples used for subsampling estimators.}

\insertNoCite{Sattler2018}{hdrm}
}
\description{
This function implements the methods outlined in
\insertCite{Sattler2018;textual}{hdrm} for data in a longtable format. For data in
a widetable format see \link{hdrm_grouped_widetable}
}
\details{
The function can deal with missing values only in \code{value}. The test is then
performed without the affected subjects. Missing values in any of the other
columns of \code{data} will result in an error.

The \code{data} is converted to a matrix and the test outlined in
\insertCite{Sattler2018;textual}{hdrm} is performed for the hypothesis given by
\code{hypothesis}. The \code{hypothesis} can either be given as a \code{character} or a \code{list}.
Legal characters are "whole" (default), "sub" or "interaction". "whole" and
"sub" stand for \eqn{H_0}: no differences in wholeplot/subplot-factor, while
"interaction" means \eqn{H_0}: no interaction between subplot- and
wholeplot-factor. Other characters will result in an error.

Alternatively, \code{hypothesis} can be a named list with quadratic
matrices \code{TW} for the wholeplot-part and \code{TS} for the subplot-part of \eqn{T
= T_W \otimes T_S}. If \code{hypothesis} is a list, \code{TW} and \code{TS} must be
idempotent, meaning symmetrical and \eqn{T_W^2 = T_W}, \eqn{T_S^2 = T_S}.
Also, the number of rows and columns of \code{TW} must be equal to the number of
groups and the number of rows and columns of \code{TS} must be equal to the number
of factor levels. Lists that do not match those criteria will cause an
error.

Note that for \code{subsampling = FALSE} your results are still seed dependent, because
the computational heaviest trace estimator is still calculated using
subsamples. Also, depending on the choice of \code{B}, the non subsampling versions might not be faster for small
data.

That also means, that even for \code{subsampling = FALSE} the \code{p.value} depends
heavily on the choice of \code{B}, as it depends on the degrees of freedom \code{f},
which are always estimated by the subsampling version. For \code{subsampling = TRUE}
the test statistic \eqn{W} is also seed dependent.

The number of subsamples \code{B} can also be a numeric value. However, this is
not advised, as \code{B} should be a function of \eqn{N} that goes to \eqn{\infty}
for \eqn{N\to\infty}.
}
\examples{
data(EEG)
?EEG
# columns of interest are:
# group = group
# value = value
# subject = subject
# factor = dimension

## Testing whether there is an effect in the diagnostic groups
hdrm_grouped_data.frame(data = EEG,
                        hypothesis = "whole",
                        group = "group",
                        value = "value",
                        subject = "subject",
                        factor = "dimension",
                        bootstrap = FALSE,
                        B = "100*N")

## columns can also be specified by their respective numbers
names(EEG)
# "group"  "value"  "sex"  "subject"  "variable"  "region"  "dimension"
hdrm_grouped_data.frame(data = EEG,
                        hypothesis = "whole",
                        group = 1,
                        value = 2,
                        subject = 4,
                        factor = 7,
                        bootstrap = FALSE,
                        B = "100*N")

## hypothesis can also be a named list with components TW and TS
## Here: TW and TS are chosen to give the same result as hypothesis = "sub"
hypothesis_list <- list(TW = matrix(1/4, 4, 4),
                        TS = diag(40) - matrix(1/40, 40, 40)
                        )
hdrm_grouped_data.frame(data = EEG,
                        hypothesis = hypothesis_list,
                        group = 1,
                        value = 2,
                        subject = 4,
                        factor = 7,
                        bootstrap = FALSE,
                        B = "100*N")
rm(hypothesis_list)
}
\references{
\insertAllCited
}
