% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_grouped_methods.R
\name{hdrm_grouped_widetable}
\alias{hdrm_grouped_widetable}
\title{Test for multiple group high dimensional repeated measures}
\usage{
hdrm_grouped_widetable(
  data,
  hypothesis = c("whole", "sub", "interaction", "identical", "flat"),
  group,
  subsampling = FALSE,
  B = "500*N",
  ...
)
}
\arguments{
\item{data}{a data frame with subjects represented by columns and factor
levels represented by rows.}

\item{hypothesis}{either one of "whole", "sub" and "interaction" or a named
list with quadratic matrices \code{TW} and \code{TS}.}

\item{group}{a \code{factor} specifying the groups.}

\item{subsampling}{\code{logical}. Specifying whether the subsampling versions for all
trace estimators should be used (see details).}

\item{B}{a \code{string} specifying a function of the number of subjects \eqn{N}.
Determines the number of subsamples used by the  subsampling trace estimators.}

\item{...}{further arguments. Currently ignored}
}
\value{
a named list of class "hdrm_grouped" with the components

\item{data}{the input data used.}

\item{f}{the estimated degrees of freedom \eqn{f}.}

\item{tau}{the convergence parameter \eqn{\tau}}

\item{H}{a named list with components \code{TW} and \code{TS} that give the
components of the hypothesis matrix}

\item{hypothesis}{a character. Will be "custom" if \code{hypothesis} is a
list, otherwise \code{hypothesis[1]}.}

\item{p.value}{the \eqn{p}-value of the test statistic.}

\item{dim}{a named list with with number of factor levels \eqn{d}
and number of subjects \eqn{N} of \code{data}}

\item{groups}{a named list with components number of groups \code{a}
and distribution of groups \code{table}.}

\item{removed.cases}{number of incomplete subjects removed.}

\item{subsamples}{number of subsamples used for subsampling estimators.}

\insertNoCite{Sattler2018}{hdrm}
}
\description{
This function implements the methods outlined
\insertCite{Sattler2018;textual}{hdrm} for data in a widetable format. For data in
a longtable format see \link{hdrm_grouped_longtable}.
}
\details{
If \code{data} contains missing values, affected subjects will be dropped with a
warning.

The test outlined in \insertCite{Sattler2018;textual}{hdrm} is performed for
the hypothesis given by \code{hypothesis}. The \code{hypothesis} can either be given as
a \code{character} or a \code{list}. Legal characters are "whole" (default), "sub" or
"interaction". "whole" and "sub" stand for \eqn{H_0}: no differences in
wholeplot/subplot-factor, while "interaction" means \eqn{H_0}: no interaction
between subplot- and wholeplot-factor. Other characters will result in an
error.

Alternatively, \code{hypothesis} can be a named list with quadratic
matrices \code{TW} for the wholeplot-part and \code{TS} for the subplot-part of \eqn{T
= T_W \otimes T_S}. If \code{hypothesis} is a list, \code{TW} and \code{TS} must be
projection matrices, meaning symmetrical and \eqn{T_W^2 = T_W}, \eqn{T_S^2 = T_S}.
Also, the number of rows and columns of \code{TW} must be equal to the number of
groups and the number of rows and columns of \code{TS} must be equal to the number
of factor levels. Lists that do not match those criteria will result in an
error.

Note that for \code{subsampling = FALSE} your results are still seed dependent, because
the computational heaviest trace estimator is still calculated using
subsamples Also, the non subsampling versions might not be faster for small
data, depending on the choice of \code{B}.

That also means, that even for \code{subsampling = FALSE} the \code{p.value} depends
heavily on the choice of \code{B}, as it depends on the estimated degrees of
freedom \code{f}, which are always estimated by the subsampling version. For
\code{subsampling = TRUE}
the test statistic \eqn{W} is also seed dependent.

The number of subsamples \code{B} can also be a numeric value. However, this is
not advised, as \code{B} should be a function of \eqn{N} that goes to \eqn{\infty}
for \eqn{N\to\infty}.
}
\examples{
data(birthrates)

# divide states into (former) east and west
group <- factor(c(1,1,2,2,1,1,1,2,1,1,1,1,2,2,1,2), labels = c("west","east"))
# call test
hdrm_grouped_widetable(data = birthrates,
                       # test for interaction effect between group and dimension
                       hypothesis = "interaction",
                       group = group,
                       bootstrap = FALSE,
                       B = "100*N"
)


# test using all subsampling version of the trace estimators
hdrm_grouped_widetable(data = birthrates,
                       # test for group effect between group and dimension
                       hypothesis = "whole",
                       group = group,
                       subsampling = TRUE,
                       # B can also be given as a number
                       B = 10000
)


# hypothesis as list
hypothesis_list <- list(TW = matrix(1/2, 2, 2),
                        TS = diag(34) - matrix(1/34, 34, 34)
) # equivalent to hypothesis = "sub"

# test call
hdrm_grouped_widetable(birthrates,
                       hypothesis = hypothesis_list,
                       group = group,
                       bootstrap = FALSE,
                       B = "100*N"
)
rm(hypothesis_list)
}
\references{
\insertAllCited
}
