% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdrm_single.R
\name{hdrm_single}
\alias{hdrm_single}
\title{Test for one group high dimensional repeated measures}
\usage{
hdrm_single(data, hypothesis = "flat", subject = NULL, AM = TRUE, ...)
}
\arguments{
\item{data}{the data for which the test is applied, where two formats are
possible:
\itemize{
\item a matrix where subjects represented by columns.
\item a value column from a data frame, where in this case an additional parameter
specifying the subjects is necessary.
}}

\item{hypothesis}{either "flat" or a quadratic numeric matrix (see Details).}

\item{subject}{optional parameter vector, which gives the allocation of the
measurements to the subjects}

\item{AM}{\code{logical}. Specifying whether an alternative hypothesis matrix
based on Sattler and Rosenbaum (2025) should be used (default: \code{TRUE}),
which has fewer rows but does not influence any values.}

\item{...}{further arguments are currently ignored.}
}
\value{
Returns a list with class "hdrm_single" with the components

\item{data}{the input data used.}

\item{f}{the degrees of freedom \eqn{f}.}

\item{statistic}{the test statistic \eqn{W}.}

\item{tau}{the convergence parameter \eqn{\tau}.}

\item{H}{the hypothesis-matrix used.}

\item{hypothesis}{a character. Will be 'custom' if \code{hypothesis} was
given as a matrix, otherwise \code{hypothesis[1]}.}

\item{p.value}{the p-value of the test statistic.}

\item{dim}{a named vector giving the dimensions \eqn{d \times N} of \code{data}.}

\item{removed.cases}{number of subjects removed for having missing values.}
}
\description{
This function implements the methods outlined in
Pauly et al. (2015).
}
\details{
If \code{data} is a matrix, the parameter \code{subject} has no effect. \cr
If \code{data} is a vector, \code{subject} allocates the measurements to the subjects.
\code{subject} then must be of length \code{length(data)}.

The function can deal with missing values only in the measurements.
The test is then performed without the affected subjects. Missing values in
\code{subject} will result in an error.

The test outlined in Pauly et al. (2015) is performed for
the hypothesis given by \code{hypothesis}. The \code{hypothesis} can either be given as
a \code{character} or a \code{list}. The only legal character is "flat" (default).
"flat" stands for \eqn{H_0}: time profile is flat. Other
characters will result in an error.

Alternatively, \code{hypothesis} can be the quadratic hypothesis matrix \eqn{T} with
the number of rows equal to the number of rows of \code{data}. \eqn{T} must be
a projection matrix, meaning symmetrical and \eqn{T^2 = T}. A matrix that does not
match those criteria will result in an error.
}
\examples{
## Load data set EEG (data frame)
data(EEG)
# ?EEG


## call test
hdrm_single(data = EEG$value,
            ## test if time profiles are flat
            hypothesis = "flat",
            ## if data is a vector, subject must be specified
            subject = EEG$subject
)


## hypothesis as list
hypothesis_matrix <- diag(40) - matrix(1/40, 40, 40)

hdrm_single(data = EEG$value,
            ## test if time profiles are flat
            hypothesis = hypothesis_matrix,
            subject = EEG$subject
)

## Load data set birthrates (matrix)
data(birthrates)
# ?birthrates

## call test
hdrm_single(data = birthrates,
            ## test if time profiles are flat
            hypothesis = "flat"
            ## if data is a matrix, subject does not need to be specified
)


## hypothesis as list
hypothesis_matrix <- diag(34) - matrix(1/34, 34, 34)

hdrm_single(data = birthrates,
            ## equivalent to hypothesis = "flat"
            hypothesis = hypothesis_matrix
)

}
\references{
{Pauly, M., Ellenberger, D., & Brunner, E. (2015). Analysis of high-dimensional one group repeated measures designs. Statistics, 49(6), 1243–1261. doi:10.1080/02331888.2015.1050022
}

{Sattler, P., & Rosenbaum, M. (2025). Choice of the hypothesis matrix for using the Anova-type-statistic. Statistics & Probability Letters, 219(110356), 110356. doi:10.1016/j.spl.2025.110356
}
}
