% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_group_case_generics.R
\name{hdrm_test}
\alias{hdrm_test}
\title{Test for multiple group high dimensional repeated measures}
\usage{
hdrm_test(
  data,
  group,
  formula,
  hypothesis = c("whole", "sub", "interaction"),
  alpha = 0.05,
  B = "500*N",
  ...
)
}
\arguments{
\item{data}{either a matrix, a list or a data.frame in longtable format containing the repeated measures data.}

\item{group}{if data is a matrix: a factor of length N that specifies groups.}

\item{formula}{a formula object that specifies the corresponding columns if data is a data.frame. Must be of the form \code{value ~ subject + group + dimension}}

\item{hypothesis}{either one of \code{whole}, \code{sub}  or \code{interaction} or a named list specifying \eqn{T_W} and \eqn{T_S}.}

\item{alpha}{the \eqn{\alpha}-level used for calculating the test statistic \eqn{W}.}

\item{B}{a string specifying a function of the number of subjects \eqn{N}. Determinates how many subsamples are used for estimating C5.}

\item{...}{further arguments. Currently ignored}
}
\value{
a named list of class "hdrm_test" with the components

\item{data}{the initial input data converted to a matrix.}

\item{f}{the degrees of freedom \eqn{f} of the critical value.}

\item{statistic}{the test statistic \eqn{W} depending on \eqn{\alpha}.}

\item{tau}{the convergence parameter \eqn{\tau}}

\item{p.value}{the \eqn{p}-value of the test statistic}

\item{critical.valie}{the critical value based on \eqn{\alpha}}

\item{dim}{a named list with components d and N of the input data}

\item{groups}{a named list with components a and table giving the number and distribution of groups}

\item{removed.cases}{number of incomplete subjects removed.}
}
\description{
Test for multiple group high dimensional repeated measures
}
\details{
If \code{data} is a matrix or a list containing multiple matrices of the same amount of rows then it is assumed, that that individuals are represented by columns and fator levels by rows.

If \code{data} is a data.frame then formula must be a formula object of the form \code{value ~ subject + group + dimension} specifying the corresponding columns of data.frame. Each subject must be represented by a unique ID.
Note, that specifying an interaction effect in  \code{formula} will give an error. For testing interaction effects, refer to \code{hypotheis}.

\code{hypothesis} can either be given as "whole" (default), "sub" or "interaction". "whole" and "sub" stand for \eqn{H_0}: no differences in wholeplot/subplot-factor, while "interaction" means \eqn{H_0}: no interaction between subplot- and wholeplot-factor.
Alternatively \code{hypothesis} can be a named list with components \code{TW} for the wholeplot-part and \code{TS} for the subplot-part of \eqn{T = T_W \otimes T_S}.
}
\references{
Paavo Sattler, Markus Pauly. "Inference for high-dimensional split-plot-designs: A unified approach for small to large numbers of factor levels." Electronic Journal of Statistics, 12(2) 2743-2805 2018. https://doi.org/10.1214/18-EJS1465
}
