% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\alias{ridge.formula}
\alias{ridge.matrix}
\title{Ridge regression}
\usage{
ridge(obj, ...)

\method{ridge}{formula}(obj, data, subset, na.action, contrasts = NULL,
  ...)

\method{ridge}{matrix}(obj, y, lambda, ...)
}
\arguments{
\item{obj}{There are two options for running `ridge()`; either supply a formula and data frame as in `lm()`, or an X and y matrix as in `glmnet()`; see examples}

\item{data}{If using the formula interface, a data.frame/list/environment containing the variables in the formula}

\item{subset}{If using the formula interface, a vector describing the subset of the data to be used in fitting the model}

\item{na.action}{If using the formula interface, a function indicating what should happen when the data contain NAs, as in `lm()`}

\item{contrasts}{If using the formula interface, a list to be passed to `model.matrix()`}

\item{y}{If using the matrix interface, the vector of responses}

\item{lam}{An optional value or vector of values for the regression coefficient; if not supplied, this will be set up automatically}
}
\description{
Ridge regression
}
\examples{
attachData(pollution)
fit <- ridge(X, y)
plot(fit)
plot(fit, xaxis='df')
plot(fit, xaxis='both')
}
\seealso{
`plot.ridge()`, `coef.ridge()`, `predict.ridge()`, `summary.ridge()`, `confint.ridge()`
}
