% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.r
\name{ridge}
\alias{ridge}
\alias{ridge.formula}
\alias{ridge.matrix}
\alias{coef.ridge}
\alias{summary.ridge}
\alias{predict.ridge}
\alias{confint.ridge}
\title{Ridge regression}
\usage{
ridge(obj, ...)

\method{ridge}{formula}(obj, data, subset, na.action, contrasts = NULL, ...)

\method{ridge}{matrix}(obj, y, lambda, ...)

\method{coef}{ridge}(
  object,
  lambda,
  which = 1:length(object$lambda),
  standardize = FALSE,
  drop = TRUE,
  ...
)

\method{summary}{ridge}(object, lambda, which, ...)

\method{predict}{ridge}(object, X, lambda, which = 1:length(object$lambda), drop = TRUE, ...)

\method{confint}{ridge}(object, parm, level = 0.95, X, lambda, which, ...)
}
\arguments{
\item{obj}{There are two options for running \code{ridge()}; either supply a formula and data frame as in \code{lm()}, or an X and y matrix as in \code{glmnet()}; see examples}

\item{...}{Additional arguments to be passed to methods}

\item{data}{If using the formula interface, a data.frame/list/environment containing the variables in the formula}

\item{subset}{If using the formula interface, a vector describing the subset of the data to be used in fitting the model}

\item{na.action}{If using the formula interface, a function indicating what should happen when the data contain NAs, as in \code{lm()}}

\item{contrasts}{If using the formula interface, a list to be passed to \code{model.matrix()}}

\item{y}{If using the matrix interface, the vector of responses}

\item{lambda}{An optional value or vector of values for the regression coefficient; if not supplied, this will be set up automatically}

\item{object}{A \code{ridge} object, as returned by \code{ridge()}}

\item{which}{Indices of \code{lambda} at which coefficients / predictions are required. By default, all
indices are returned. If \code{lambda} is specified, this will override \code{which}.}

\item{standardize}{Return coefficients on standardized scale (default: FALSE)}

\item{drop}{If requesting coefficients for a single lambda value, drop matrix down to a vector (default: TRUE)}

\item{X}{Matrix of predictor values at which predictions are required.}

\item{parm}{Which parameters to construct confidence intervals for; either a vector of indices or a vector of names (default: all)}

\item{level}{Confidence level (default: 0.95)}
}
\description{
Ridge regression
}
\examples{
attachData(pollution)
fit <- ridge(X, y)
summary(fit, lambda=0.1)
plot(fit)
plot(fit, xaxis='df')
plot(fit, xaxis='both')
plot(fit, standardize=TRUE)
coef(fit)
coef(fit, standardize=TRUE)
coef(fit, standardize=TRUE, lambda=0.1)
p1 <- cbind(1, std(X)) \%*\% coef(fit, standardize=TRUE, lambda=1)
p2 <- cbind(1, X) \%*\% coef(fit, lambda=1)
head(p1); head(p2) # identical
}
\seealso{
\code{\link[=plot.ridge]{plot.ridge()}}, \code{coef.ridge()}, \code{predict.ridge()}, \code{summary.ridge()}, \code{confint.ridge()}
}
